@@ -1,5 +1,5 @@
-// flow-typed signature: 74fc11ed78023f391f225b50e2b311a4
-// flow-typed version: <<STUB>>/@babel/core_v^7.4.5/flow_v0.99.1
+// flow-typed signature: 23eb20a8b396fbcaa5dc18ed975a035e
+// flow-typed version: <<STUB>>/@babel/core_v^7.7.2/flow_v0.112.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -42,7 +42,7 @@ declare module '@babel/core/lib/config/files/index-browser' {
   declare module.exports: any;
 }
 
-declare module '@babel/core/lib/config/files/index' {
+declare module '@babel/core/lib/config/files' {
   declare module.exports: any;
 }
 
@@ -74,7 +74,7 @@ declare module '@babel/core/lib/config/helpers/environment' {
   declare module.exports: any;
 }
 
-declare module '@babel/core/lib/config/index' {
+declare module '@babel/core/lib/config' {
   declare module.exports: any;
 }
 
@@ -114,7 +114,7 @@ declare module '@babel/core/lib/config/validation/removed' {
   declare module.exports: any;
 }
 
-declare module '@babel/core/lib/index' {
+declare module '@babel/core/lib' {
   declare module.exports: any;
 }
 
@@ -158,7 +158,7 @@ declare module '@babel/core/lib/transformation/file/merge-map' {
   declare module.exports: any;
 }
 
-declare module '@babel/core/lib/transformation/index' {
+declare module '@babel/core/lib/transformation' {
   declare module.exports: any;
 }
 
@@ -189,13 +189,20 @@ declare module '@babel/core/lib/config/config-descriptors.js' {
   declare module.exports: $Exports<'@babel/core/lib/config/config-descriptors'>;
 }
 declare module '@babel/core/lib/config/files/configuration.js' {
-  declare module.exports: $Exports<'@babel/core/lib/config/files/configuration'>;
+  declare module.exports: $Exports<
+    '@babel/core/lib/config/files/configuration'
+  >;
 }
 declare module '@babel/core/lib/config/files/index-browser.js' {
-  declare module.exports: $Exports<'@babel/core/lib/config/files/index-browser'>;
+  declare module.exports: $Exports<
+    '@babel/core/lib/config/files/index-browser'
+  >;
+}
+declare module '@babel/core/lib/config/files/index' {
+  declare module.exports: $Exports<'@babel/core/lib/config/files'>;
 }
 declare module '@babel/core/lib/config/files/index.js' {
-  declare module.exports: $Exports<'@babel/core/lib/config/files/index'>;
+  declare module.exports: $Exports<'@babel/core/lib/config/files'>;
 }
 declare module '@babel/core/lib/config/files/package.js' {
   declare module.exports: $Exports<'@babel/core/lib/config/files/package'>;
@@ -216,10 +223,15 @@ declare module '@babel/core/lib/config/helpers/config-api.js' {
   declare module.exports: $Exports<'@babel/core/lib/config/helpers/config-api'>;
 }
 declare module '@babel/core/lib/config/helpers/environment.js' {
-  declare module.exports: $Exports<'@babel/core/lib/config/helpers/environment'>;
+  declare module.exports: $Exports<
+    '@babel/core/lib/config/helpers/environment'
+  >;
+}
+declare module '@babel/core/lib/config/index' {
+  declare module.exports: $Exports<'@babel/core/lib/config'>;
 }
 declare module '@babel/core/lib/config/index.js' {
-  declare module.exports: $Exports<'@babel/core/lib/config/index'>;
+  declare module.exports: $Exports<'@babel/core/lib/config'>;
 }
 declare module '@babel/core/lib/config/item.js' {
   declare module.exports: $Exports<'@babel/core/lib/config/item'>;
@@ -237,7 +249,9 @@ declare module '@babel/core/lib/config/util.js' {
   declare module.exports: $Exports<'@babel/core/lib/config/util'>;
 }
 declare module '@babel/core/lib/config/validation/option-assertions.js' {
-  declare module.exports: $Exports<'@babel/core/lib/config/validation/option-assertions'>;
+  declare module.exports: $Exports<
+    '@babel/core/lib/config/validation/option-assertions'
+  >;
 }
 declare module '@babel/core/lib/config/validation/options.js' {
   declare module.exports: $Exports<'@babel/core/lib/config/validation/options'>;
@@ -248,14 +262,19 @@ declare module '@babel/core/lib/config/validation/plugins.js' {
 declare module '@babel/core/lib/config/validation/removed.js' {
   declare module.exports: $Exports<'@babel/core/lib/config/validation/removed'>;
 }
+declare module '@babel/core/lib/index' {
+  declare module.exports: $Exports<'@babel/core/lib'>;
+}
 declare module '@babel/core/lib/index.js' {
-  declare module.exports: $Exports<'@babel/core/lib/index'>;
+  declare module.exports: $Exports<'@babel/core/lib'>;
 }
 declare module '@babel/core/lib/parse.js' {
   declare module.exports: $Exports<'@babel/core/lib/parse'>;
 }
 declare module '@babel/core/lib/tools/build-external-helpers.js' {
-  declare module.exports: $Exports<'@babel/core/lib/tools/build-external-helpers'>;
+  declare module.exports: $Exports<
+    '@babel/core/lib/tools/build-external-helpers'
+  >;
 }
 declare module '@babel/core/lib/transform-ast.js' {
   declare module.exports: $Exports<'@babel/core/lib/transform-ast'>;
@@ -270,29 +289,46 @@ declare module '@babel/core/lib/transform.js' {
   declare module.exports: $Exports<'@babel/core/lib/transform'>;
 }
 declare module '@babel/core/lib/transformation/block-hoist-plugin.js' {
-  declare module.exports: $Exports<'@babel/core/lib/transformation/block-hoist-plugin'>;
+  declare module.exports: $Exports<
+    '@babel/core/lib/transformation/block-hoist-plugin'
+  >;
 }
 declare module '@babel/core/lib/transformation/file/file.js' {
   declare module.exports: $Exports<'@babel/core/lib/transformation/file/file'>;
 }
 declare module '@babel/core/lib/transformation/file/generate.js' {
-  declare module.exports: $Exports<'@babel/core/lib/transformation/file/generate'>;
+  declare module.exports: $Exports<
+    '@babel/core/lib/transformation/file/generate'
+  >;
 }
 declare module '@babel/core/lib/transformation/file/merge-map.js' {
-  declare module.exports: $Exports<'@babel/core/lib/transformation/file/merge-map'>;
+  declare module.exports: $Exports<
+    '@babel/core/lib/transformation/file/merge-map'
+  >;
+}
+declare module '@babel/core/lib/transformation/index' {
+  declare module.exports: $Exports<'@babel/core/lib/transformation'>;
 }
 declare module '@babel/core/lib/transformation/index.js' {
-  declare module.exports: $Exports<'@babel/core/lib/transformation/index'>;
+  declare module.exports: $Exports<'@babel/core/lib/transformation'>;
 }
 declare module '@babel/core/lib/transformation/normalize-file.js' {
-  declare module.exports: $Exports<'@babel/core/lib/transformation/normalize-file'>;
+  declare module.exports: $Exports<
+    '@babel/core/lib/transformation/normalize-file'
+  >;
 }
 declare module '@babel/core/lib/transformation/normalize-opts.js' {
-  declare module.exports: $Exports<'@babel/core/lib/transformation/normalize-opts'>;
+  declare module.exports: $Exports<
+    '@babel/core/lib/transformation/normalize-opts'
+  >;
 }
 declare module '@babel/core/lib/transformation/plugin-pass.js' {
-  declare module.exports: $Exports<'@babel/core/lib/transformation/plugin-pass'>;
+  declare module.exports: $Exports<
+    '@babel/core/lib/transformation/plugin-pass'
+  >;
 }
 declare module '@babel/core/lib/transformation/util/missing-plugin-helper.js' {
-  declare module.exports: $Exports<'@babel/core/lib/transformation/util/missing-plugin-helper'>;
+  declare module.exports: $Exports<
+    '@babel/core/lib/transformation/util/missing-plugin-helper'
+  >;
 }
