@@ -1,5 +1,5 @@
-// flow-typed signature: 0a4a767571b1c51d76b02eb1c21defb8
-// flow-typed version: <<STUB>>/@babel/preset-flow_v7.0.0/flow_v0.99.1
+// flow-typed signature: e25575fb9e2dbf3863f9bfd2909147d9
+// flow-typed version: <<STUB>>/@babel/preset-flow_v7.0.0/flow_v0.112.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -22,11 +22,14 @@ declare module '@babel/preset-flow' {
  * require those files directly. Feel free to delete any files that aren't
  * needed.
  */
-declare module '@babel/preset-flow/lib/index' {
+declare module '@babel/preset-flow/lib' {
   declare module.exports: any;
 }
 
 // Filename aliases
+declare module '@babel/preset-flow/lib/index' {
+  declare module.exports: $Exports<'@babel/preset-flow/lib'>;
+}
 declare module '@babel/preset-flow/lib/index.js' {
-  declare module.exports: $Exports<'@babel/preset-flow/lib/index'>;
+  declare module.exports: $Exports<'@babel/preset-flow/lib'>;
 }
