@@ -1,7 +1,19 @@
-// flow-typed signature: 59eda4ec84d59ec5e2a7ba8354c5f45b
-// flow-typed version: ba34c6466e/react-testing-library_v5.x.x/flow_>=v0.67.1
+// flow-typed signature: e0056fd09a33a332c0ad120f7d50308d
+// flow-typed version: 452cd002ef/@testing-library/react_v9.x.x/flow_>=v0.104.x
 
-declare module 'react-testing-library' {
+declare module '@testing-library/react' {
+  // This type comes from
+  // https://github.com/facebook/flow/blob/v0.104.0/lib/react-dom.js#L64
+  declare type ReactDOMTestUtilsThenable = {
+    then(resolve: () => mixed, reject?: () => mixed): mixed,
+    ...
+  };
+  // This type comes from
+  // https://github.com/facebook/flow/blob/v0.104.0/lib/react-dom.js#L116
+  declare type ReactDOMTestUtilsAct = (
+    callback: () => void | ReactDOMTestUtilsThenable
+  ) => ReactDOMTestUtilsThenable;
+
   declare type TextMatch =
     | string
     | RegExp
@@ -11,87 +23,160 @@ declare module 'react-testing-library' {
     exact?: boolean,
     trim?: boolean,
     collapseWhitespace?: boolean,
+    ...
   };
 
-  declare type GetsAndQueries = {|
-    getByTestId: (id: TextMatch, options?: TextMatchOptions) => HTMLElement,
-    getByText: (
-      text: TextMatch,
-      options?: { selector?: string } & TextMatchOptions
-    ) => HTMLElement,
-    getByPlaceholderText: (
-      text: TextMatch,
-      options?: TextMatchOptions
-    ) => HTMLElement,
-    getByLabelText: (
-      text: TextMatch,
-      options?: { selector?: string } & TextMatchOptions
-    ) => HTMLElement,
-    getByAltText: (text: TextMatch, options?: TextMatchOptions) => HTMLElement,
-    getAll: (text: TextMatch, options?: TextMatchOptions) => Array<HTMLElement>,
-    queryByTestId: (id: TextMatch, options?: TextMatchOptions) => ?HTMLElement,
-    queryByText: (text: TextMatch, options?: TextMatchOptions) => ?HTMLElement,
-    queryByPlaceholderText: (
-      text: TextMatch,
-      options?: TextMatchOptions
-    ) => ?HTMLElement,
-    queryByLabelText: (
-      text: TextMatch,
-      options?: TextMatchOptions
-    ) => ?HTMLElement,
-    queryByAltText: (
-      text: TextMatch,
-      options?: TextMatchOptions
-    ) => ?HTMLElement,
-    queryAll: (
-      text: TextMatch,
-      options?: TextMatchOptions
-    ) => Array<HTMLElement>,
-  |};
+  declare type SelectorMatchOptions = {
+    selector?: string,
+    ...
+  } & TextMatchOptions;
 
-  declare type RenderResult = {|
-    container: HTMLDivElement,
-    unmount: () => void,
-    baseElement: HTMLElement,
-    debug: (baseElement?: HTMLElement) => void,
-    rerender: (ui: React$Element<*>) => void,
-  |} & GetsAndQueries;
+  declare type GetByText = (
+    text: TextMatch,
+    options?: SelectorMatchOptions
+  ) => HTMLElement;
+
+  declare type QueryByText = (
+    text: TextMatch,
+    options?: SelectorMatchOptions
+  ) => ?HTMLElement;
+
+  declare type AllByText = (
+    text: TextMatch,
+    options?: SelectorMatchOptions
+  ) => Array<HTMLElement>;
+
+  declare type GetByBoundAttribute = (
+    text: TextMatch,
+    options?: TextMatchOptions
+  ) => HTMLElement;
+
+  declare type QueryByBoundAttribute = (
+    text: TextMatch,
+    options?: TextMatchOptions
+  ) => ?HTMLElement;
+
+  declare type AllByBoundAttribute = (
+    text: TextMatch,
+    options?: TextMatchOptions
+  ) => Array<HTMLElement>;
+
+  declare type GetsAndQueries = {|
+    getByAltText: GetByBoundAttribute,
+    getAllByAltText: AllByBoundAttribute,
+    queryByAltText: QueryByBoundAttribute,
+    queryAllByAltText: AllByBoundAttribute,
+
+    getByDisplayValue: GetByBoundAttribute,
+    getAllByDisplayValue: AllByBoundAttribute,
+    queryByDisplayValue: QueryByBoundAttribute,
+    queryAllByDisplayValue: AllByBoundAttribute,
+
+    getByLabelText: GetByText,
+    getAllByLabelText: AllByText,
+    queryByLabelText: QueryByText,
+    queryAllByLabelText: AllByText,
+
+    getByPlaceholderText: GetByBoundAttribute,
+    getAllByPlaceholderText: AllByBoundAttribute,
+    queryByPlaceholderText: QueryByBoundAttribute,
+    queryAllByPlaceholderText: AllByBoundAttribute,
+
+    getByRole: GetByBoundAttribute,
+    getAllByRole: AllByBoundAttribute,
+    queryByRole: QueryByBoundAttribute,
+    queryAllByRole: AllByBoundAttribute,
+
+    getBySelectText: GetByBoundAttribute,
+    getAllBySelectText: AllByBoundAttribute,
+    queryBySelectText: QueryByBoundAttribute,
+    queryAllBySelectText: AllByBoundAttribute,
+
+    getByTestId: GetByBoundAttribute,
+    getAllByTestId: AllByBoundAttribute,
+    queryByTestId: QueryByBoundAttribute,
+    queryAllByTestId: AllByBoundAttribute,
+
+    getByText: GetByText,
+    getAllByText: AllByText,
+    queryByText: QueryByText,
+    queryAllByText: AllByText,
+
+    getByTitle: GetByBoundAttribute,
+    getAllByTitle: AllByBoundAttribute,
+    queryByTitle: QueryByBoundAttribute,
+    queryAllByTitle: AllByBoundAttribute,
+
+    getByValue: GetByBoundAttribute,
+    getAllByValue: AllByBoundAttribute,
+    queryByValue: QueryByBoundAttribute,
+    queryAllByValue: AllByBoundAttribute,
+  |};
 
   declare type FireEvent<TInit> = (
     element: HTMLElement,
     eventProperties?: TInit
   ) => boolean;
 
+  declare type Queries = { ... };
+
+  declare type RenderResult<Q: Queries = GetsAndQueries> = {|
+    container: HTMLDivElement,
+    unmount: () => void,
+    baseElement: HTMLElement,
+    asFragment: () => DocumentFragment,
+    debug: (baseElement?: HTMLElement) => void,
+    rerender: (ui: React$Element<*>) => void,
+  |} & Q;
+
+  declare export type RenderOptions<Q: Queries = { ... }> = {|
+    container?: HTMLElement,
+    baseElement?: HTMLElement,
+    hydrate?: boolean,
+    queries?: Q,
+    wrapper?: React.ComponentType,
+  |};
+
   declare module.exports: {
-    render: (
-      ui: React$Element<*>,
-      options?: { container: HTMLElement, baseElement?: HTMLElement }
-    ) => RenderResult,
+    render(
+      ui: React.ReactElement<any>,
+      options?: $Diff<RenderOptions<>, {| queries: any |}>
+    ): RenderResult<>,
 
+    render<Q: Queries>(
+      ui: React.ReactElement<any>,
+      options?: RenderOptions<Q>
+    ): RenderResult<Q>,
+
+    act: ReactDOMTestUtilsAct,
     cleanup: () => void,
-
     wait: (
       callback?: () => void,
       options?: {
         timeout?: number,
         interval?: number,
+        ...
       }
     ) => Promise<void>,
-
+    waitForDomChange: <T>(options?: {
+      container?: HTMLElement,
+      timeout?: number,
+      mutationObserverOptions?: MutationObserverInit,
+      ...
+    }) => Promise<T>,
     waitForElement: <T>(
       callback?: () => T,
       options?: {
         container?: HTMLElement,
         timeout?: number,
         mutationObserverOptions?: MutationObserverInit,
+        ...
       }
-    ) => Promise<T | void>,
-
+    ) => Promise<T>,
     within: (
       element: HTMLElement,
       queriesToBind?: GetsAndQueries | Array<GetsAndQueries>
     ) => GetsAndQueries,
-
     fireEvent: {|
       (element: HTMLElement, event: Event): void,
 
@@ -165,7 +250,6 @@ declare module 'react-testing-library' {
       animationIteration: FireEvent<Event$Init>,
       transitionEnd: FireEvent<Event$Init>,
     |},
-
     // dom-testing-library re-exports
     queryByTestId: (
       container: HTMLElement,
@@ -185,7 +269,7 @@ declare module 'react-testing-library' {
     getByText: (
       container: HTMLElement,
       text: TextMatch,
-      options?: { selector?: string } & TextMatchOptions
+      options?: { selector?: string, ... } & TextMatchOptions
     ) => HTMLElement,
     queryByPlaceholderText: (
       container: HTMLElement,
@@ -205,7 +289,7 @@ declare module 'react-testing-library' {
     getByLabelText: (
       container: HTMLElement,
       text: TextMatch,
-      options?: { selector?: string } & TextMatchOptions
+      options?: { selector?: string, ... } & TextMatchOptions
     ) => HTMLElement,
     queryByAltText: (
       container: HTMLElement,
@@ -217,5 +301,6 @@ declare module 'react-testing-library' {
       text: TextMatch,
       options?: TextMatchOptions
     ) => HTMLElement,
+    ...
   };
 }
