@@ -1,5 +1,5 @@
-// flow-typed signature: 1fcca4a830c671309b229406d1380801
-// flow-typed version: <<STUB>>/babel-core_v7.0.0-bridge.0/flow_v0.99.1
+// flow-typed signature: d878a7d0a1571dced1e2e6ab1a9ba0e5
+// flow-typed version: <<STUB>>/babel-core_v7.0.0-bridge.0/flow_v0.112.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -22,7 +22,6 @@ declare module 'babel-core' {
  * require those files directly. Feel free to delete any files that aren't
  * needed.
  */
-
 
 // Filename aliases
 declare module 'babel-core/index' {
