@@ -1,5 +1,5 @@
-// flow-typed signature: 036d3b29f52af143bf7e7c7fca98b9f6
-// flow-typed version: <<STUB>>/babel-eslint_v10.0.1/flow_v0.99.1
+// flow-typed signature: e3cee9dfee0ad0c03de73fcff4acd4c2
+// flow-typed version: <<STUB>>/babel-eslint_v10.0.3/flow_v0.112.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -38,7 +38,7 @@ declare module 'babel-eslint/lib/babylon-to-espree/convertTemplateType' {
   declare module.exports: any;
 }
 
-declare module 'babel-eslint/lib/babylon-to-espree/index' {
+declare module 'babel-eslint/lib/babylon-to-espree' {
   declare module.exports: any;
 }
 
@@ -54,7 +54,7 @@ declare module 'babel-eslint/lib/babylon-to-espree/toTokens' {
   declare module.exports: any;
 }
 
-declare module 'babel-eslint/lib/index' {
+declare module 'babel-eslint/lib' {
   declare module.exports: any;
 }
 
@@ -63,6 +63,10 @@ declare module 'babel-eslint/lib/parse-with-scope' {
 }
 
 declare module 'babel-eslint/lib/parse' {
+  declare module.exports: any;
+}
+
+declare module 'babel-eslint/lib/require-from-eslint' {
   declare module.exports: any;
 }
 
@@ -75,34 +79,53 @@ declare module 'babel-eslint/lib/analyze-scope.js' {
   declare module.exports: $Exports<'babel-eslint/lib/analyze-scope'>;
 }
 declare module 'babel-eslint/lib/babylon-to-espree/attachComments.js' {
-  declare module.exports: $Exports<'babel-eslint/lib/babylon-to-espree/attachComments'>;
+  declare module.exports: $Exports<
+    'babel-eslint/lib/babylon-to-espree/attachComments'
+  >;
 }
 declare module 'babel-eslint/lib/babylon-to-espree/convertComments.js' {
-  declare module.exports: $Exports<'babel-eslint/lib/babylon-to-espree/convertComments'>;
+  declare module.exports: $Exports<
+    'babel-eslint/lib/babylon-to-espree/convertComments'
+  >;
 }
 declare module 'babel-eslint/lib/babylon-to-espree/convertTemplateType.js' {
-  declare module.exports: $Exports<'babel-eslint/lib/babylon-to-espree/convertTemplateType'>;
+  declare module.exports: $Exports<
+    'babel-eslint/lib/babylon-to-espree/convertTemplateType'
+  >;
+}
+declare module 'babel-eslint/lib/babylon-to-espree/index' {
+  declare module.exports: $Exports<'babel-eslint/lib/babylon-to-espree'>;
 }
 declare module 'babel-eslint/lib/babylon-to-espree/index.js' {
-  declare module.exports: $Exports<'babel-eslint/lib/babylon-to-espree/index'>;
+  declare module.exports: $Exports<'babel-eslint/lib/babylon-to-espree'>;
 }
 declare module 'babel-eslint/lib/babylon-to-espree/toAST.js' {
   declare module.exports: $Exports<'babel-eslint/lib/babylon-to-espree/toAST'>;
 }
 declare module 'babel-eslint/lib/babylon-to-espree/toToken.js' {
-  declare module.exports: $Exports<'babel-eslint/lib/babylon-to-espree/toToken'>;
+  declare module.exports: $Exports<
+    'babel-eslint/lib/babylon-to-espree/toToken'
+  >;
 }
 declare module 'babel-eslint/lib/babylon-to-espree/toTokens.js' {
-  declare module.exports: $Exports<'babel-eslint/lib/babylon-to-espree/toTokens'>;
+  declare module.exports: $Exports<
+    'babel-eslint/lib/babylon-to-espree/toTokens'
+  >;
+}
+declare module 'babel-eslint/lib/index' {
+  declare module.exports: $Exports<'babel-eslint/lib'>;
 }
 declare module 'babel-eslint/lib/index.js' {
-  declare module.exports: $Exports<'babel-eslint/lib/index'>;
+  declare module.exports: $Exports<'babel-eslint/lib'>;
 }
 declare module 'babel-eslint/lib/parse-with-scope.js' {
   declare module.exports: $Exports<'babel-eslint/lib/parse-with-scope'>;
 }
 declare module 'babel-eslint/lib/parse.js' {
   declare module.exports: $Exports<'babel-eslint/lib/parse'>;
+}
+declare module 'babel-eslint/lib/require-from-eslint.js' {
+  declare module.exports: $Exports<'babel-eslint/lib/require-from-eslint'>;
 }
 declare module 'babel-eslint/lib/visitor-keys.js' {
   declare module.exports: $Exports<'babel-eslint/lib/visitor-keys'>;
