@@ -1,5 +1,5 @@
-// flow-typed signature: 6ba3ab8ea12728a5d7b3d3ba29177371
-// flow-typed version: <<STUB>>/babel-jest_v24.8.0/flow_v0.99.1
+// flow-typed signature: ad623e8f15f453902fa325fbb391788f
+// flow-typed version: <<STUB>>/babel-jest_v24.9.0/flow_v0.112.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -22,11 +22,14 @@ declare module 'babel-jest' {
  * require those files directly. Feel free to delete any files that aren't
  * needed.
  */
-declare module 'babel-jest/build/index' {
+declare module 'babel-jest/build' {
   declare module.exports: any;
 }
 
 // Filename aliases
+declare module 'babel-jest/build/index' {
+  declare module.exports: $Exports<'babel-jest/build'>;
+}
 declare module 'babel-jest/build/index.js' {
-  declare module.exports: $Exports<'babel-jest/build/index'>;
+  declare module.exports: $Exports<'babel-jest/build'>;
 }
