@@ -1,5 +1,5 @@
-// flow-typed signature: e0fde7ff7c7fe784fa3f09d034f1f5f2
-// flow-typed version: <<STUB>>/babel-plugin-styled-components_v1.10.0/flow_v0.99.1
+// flow-typed signature: 58868a06d5aa21bf8ad745810a58fb2b
+// flow-typed version: <<STUB>>/babel-plugin-styled-components_v1.10.6/flow_v0.112.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -26,11 +26,11 @@ declare module 'babel-plugin-styled-components/lib/css/placeholderUtils' {
   declare module.exports: any;
 }
 
-declare module 'babel-plugin-styled-components/lib/index' {
+declare module 'babel-plugin-styled-components/lib' {
   declare module.exports: any;
 }
 
-declare module 'babel-plugin-styled-components/lib/minify/index' {
+declare module 'babel-plugin-styled-components/lib/minify' {
   declare module.exports: any;
 }
 
@@ -70,7 +70,7 @@ declare module 'babel-plugin-styled-components/lib/visitors/pure' {
   declare module.exports: any;
 }
 
-declare module 'babel-plugin-styled-components/lib/visitors/templateLiterals/index' {
+declare module 'babel-plugin-styled-components/lib/visitors/templateLiterals' {
   declare module.exports: any;
 }
 
@@ -84,47 +84,84 @@ declare module 'babel-plugin-styled-components/lib/visitors/transpileCssProp' {
 
 // Filename aliases
 declare module 'babel-plugin-styled-components/lib/css/placeholderUtils.js' {
-  declare module.exports: $Exports<'babel-plugin-styled-components/lib/css/placeholderUtils'>;
+  declare module.exports: $Exports<
+    'babel-plugin-styled-components/lib/css/placeholderUtils'
+  >;
+}
+declare module 'babel-plugin-styled-components/lib/index' {
+  declare module.exports: $Exports<'babel-plugin-styled-components/lib'>;
 }
 declare module 'babel-plugin-styled-components/lib/index.js' {
-  declare module.exports: $Exports<'babel-plugin-styled-components/lib/index'>;
+  declare module.exports: $Exports<'babel-plugin-styled-components/lib'>;
+}
+declare module 'babel-plugin-styled-components/lib/minify/index' {
+  declare module.exports: $Exports<'babel-plugin-styled-components/lib/minify'>;
 }
 declare module 'babel-plugin-styled-components/lib/minify/index.js' {
-  declare module.exports: $Exports<'babel-plugin-styled-components/lib/minify/index'>;
+  declare module.exports: $Exports<'babel-plugin-styled-components/lib/minify'>;
 }
 declare module 'babel-plugin-styled-components/lib/utils/detectors.js' {
-  declare module.exports: $Exports<'babel-plugin-styled-components/lib/utils/detectors'>;
+  declare module.exports: $Exports<
+    'babel-plugin-styled-components/lib/utils/detectors'
+  >;
 }
 declare module 'babel-plugin-styled-components/lib/utils/getName.js' {
-  declare module.exports: $Exports<'babel-plugin-styled-components/lib/utils/getName'>;
+  declare module.exports: $Exports<
+    'babel-plugin-styled-components/lib/utils/getName'
+  >;
 }
 declare module 'babel-plugin-styled-components/lib/utils/hash.js' {
-  declare module.exports: $Exports<'babel-plugin-styled-components/lib/utils/hash'>;
+  declare module.exports: $Exports<
+    'babel-plugin-styled-components/lib/utils/hash'
+  >;
 }
 declare module 'babel-plugin-styled-components/lib/utils/options.js' {
-  declare module.exports: $Exports<'babel-plugin-styled-components/lib/utils/options'>;
+  declare module.exports: $Exports<
+    'babel-plugin-styled-components/lib/utils/options'
+  >;
 }
 declare module 'babel-plugin-styled-components/lib/utils/prefixDigit.js' {
-  declare module.exports: $Exports<'babel-plugin-styled-components/lib/utils/prefixDigit'>;
+  declare module.exports: $Exports<
+    'babel-plugin-styled-components/lib/utils/prefixDigit'
+  >;
 }
 declare module 'babel-plugin-styled-components/lib/visitors/assignStyledRequired.js' {
-  declare module.exports: $Exports<'babel-plugin-styled-components/lib/visitors/assignStyledRequired'>;
+  declare module.exports: $Exports<
+    'babel-plugin-styled-components/lib/visitors/assignStyledRequired'
+  >;
 }
 declare module 'babel-plugin-styled-components/lib/visitors/displayNameAndId.js' {
-  declare module.exports: $Exports<'babel-plugin-styled-components/lib/visitors/displayNameAndId'>;
+  declare module.exports: $Exports<
+    'babel-plugin-styled-components/lib/visitors/displayNameAndId'
+  >;
 }
 declare module 'babel-plugin-styled-components/lib/visitors/minify.js' {
-  declare module.exports: $Exports<'babel-plugin-styled-components/lib/visitors/minify'>;
+  declare module.exports: $Exports<
+    'babel-plugin-styled-components/lib/visitors/minify'
+  >;
 }
 declare module 'babel-plugin-styled-components/lib/visitors/pure.js' {
-  declare module.exports: $Exports<'babel-plugin-styled-components/lib/visitors/pure'>;
+  declare module.exports: $Exports<
+    'babel-plugin-styled-components/lib/visitors/pure'
+  >;
+}
+declare module 'babel-plugin-styled-components/lib/visitors/templateLiterals/index' {
+  declare module.exports: $Exports<
+    'babel-plugin-styled-components/lib/visitors/templateLiterals'
+  >;
 }
 declare module 'babel-plugin-styled-components/lib/visitors/templateLiterals/index.js' {
-  declare module.exports: $Exports<'babel-plugin-styled-components/lib/visitors/templateLiterals/index'>;
+  declare module.exports: $Exports<
+    'babel-plugin-styled-components/lib/visitors/templateLiterals'
+  >;
 }
 declare module 'babel-plugin-styled-components/lib/visitors/templateLiterals/transpile.js' {
-  declare module.exports: $Exports<'babel-plugin-styled-components/lib/visitors/templateLiterals/transpile'>;
+  declare module.exports: $Exports<
+    'babel-plugin-styled-components/lib/visitors/templateLiterals/transpile'
+  >;
 }
 declare module 'babel-plugin-styled-components/lib/visitors/transpileCssProp.js' {
-  declare module.exports: $Exports<'babel-plugin-styled-components/lib/visitors/transpileCssProp'>;
+  declare module.exports: $Exports<
+    'babel-plugin-styled-components/lib/visitors/transpileCssProp'
+  >;
 }
