@@ -1,5 +1,5 @@
-// flow-typed signature: fda6e43211bc2a9ec4aad433dc490506
-// flow-typed version: <<STUB>>/babel-plugin-transform-dynamic-import_v2.1.0/flow_v0.99.1
+// flow-typed signature: ad9ce2a70509e87402c9968a125aead1
+// flow-typed version: <<STUB>>/babel-plugin-transform-dynamic-import_v2.1.0/flow_v0.112.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -22,11 +22,14 @@ declare module 'babel-plugin-transform-dynamic-import' {
  * require those files directly. Feel free to delete any files that aren't
  * needed.
  */
-declare module 'babel-plugin-transform-dynamic-import/lib/index' {
+declare module 'babel-plugin-transform-dynamic-import/lib' {
   declare module.exports: any;
 }
 
 // Filename aliases
+declare module 'babel-plugin-transform-dynamic-import/lib/index' {
+  declare module.exports: $Exports<'babel-plugin-transform-dynamic-import/lib'>;
+}
 declare module 'babel-plugin-transform-dynamic-import/lib/index.js' {
-  declare module.exports: $Exports<'babel-plugin-transform-dynamic-import/lib/index'>;
+  declare module.exports: $Exports<'babel-plugin-transform-dynamic-import/lib'>;
 }
