@@ -1,5 +1,5 @@
-// flow-typed signature: bd5af9d412166a08e04b4d8b01e4bafd
-// flow-typed version: <<STUB>>/eslint-config-airbnb_v17.1.0/flow_v0.99.1
+// flow-typed signature: 4bf87240ce2c8dc65ac75db092f91e80
+// flow-typed version: <<STUB>>/eslint-config-airbnb_v18.0.1/flow_v0.112.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -26,11 +26,19 @@ declare module 'eslint-config-airbnb/base' {
   declare module.exports: any;
 }
 
+declare module 'eslint-config-airbnb/hooks' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-config-airbnb/legacy' {
   declare module.exports: any;
 }
 
 declare module 'eslint-config-airbnb/rules/react-a11y' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-config-airbnb/rules/react-hooks' {
   declare module.exports: any;
 }
 
@@ -58,6 +66,9 @@ declare module 'eslint-config-airbnb/whitespace' {
 declare module 'eslint-config-airbnb/base.js' {
   declare module.exports: $Exports<'eslint-config-airbnb/base'>;
 }
+declare module 'eslint-config-airbnb/hooks.js' {
+  declare module.exports: $Exports<'eslint-config-airbnb/hooks'>;
+}
 declare module 'eslint-config-airbnb/index' {
   declare module.exports: $Exports<'eslint-config-airbnb'>;
 }
@@ -70,6 +81,9 @@ declare module 'eslint-config-airbnb/legacy.js' {
 declare module 'eslint-config-airbnb/rules/react-a11y.js' {
   declare module.exports: $Exports<'eslint-config-airbnb/rules/react-a11y'>;
 }
+declare module 'eslint-config-airbnb/rules/react-hooks.js' {
+  declare module.exports: $Exports<'eslint-config-airbnb/rules/react-hooks'>;
+}
 declare module 'eslint-config-airbnb/rules/react.js' {
   declare module.exports: $Exports<'eslint-config-airbnb/rules/react'>;
 }
@@ -80,7 +94,9 @@ declare module 'eslint-config-airbnb/test/test-base.js' {
   declare module.exports: $Exports<'eslint-config-airbnb/test/test-base'>;
 }
 declare module 'eslint-config-airbnb/test/test-react-order.js' {
-  declare module.exports: $Exports<'eslint-config-airbnb/test/test-react-order'>;
+  declare module.exports: $Exports<
+    'eslint-config-airbnb/test/test-react-order'
+  >;
 }
 declare module 'eslint-config-airbnb/whitespace.js' {
   declare module.exports: $Exports<'eslint-config-airbnb/whitespace'>;
