@@ -1,5 +1,5 @@
-// flow-typed signature: 2a2ace9a04cdb1b81d0a275fb0f85852
-// flow-typed version: <<STUB>>/eslint-config-prettier_v4.3.0/flow_v0.99.1
+// flow-typed signature: 7f5bd153f4c58c0385541f26ee3e9c77
+// flow-typed version: <<STUB>>/eslint-config-prettier_v6.6.0/flow_v0.112.0
 
 /**
  * This is an autogenerated libdef stub for:
