@@ -1,5 +1,5 @@
-// flow-typed signature: 2527fc0cf5b4ffc79a1f3b132e98702f
-// flow-typed version: <<STUB>>/eslint-plugin-import_v2.17.3/flow_v0.99.1
+// flow-typed signature: 35d1f5958822a3d21d13ca4846f30840
+// flow-typed version: <<STUB>>/eslint-plugin-import_v2.18.2/flow_v0.112.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -74,7 +74,7 @@ declare module 'eslint-plugin-import/lib/importDeclaration' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-import/lib/index' {
+declare module 'eslint-plugin-import/lib' {
   declare module.exports: any;
 }
 
@@ -242,7 +242,7 @@ declare module 'eslint-plugin-import/lib/rules/unambiguous' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-import/memo-parser/index' {
+declare module 'eslint-plugin-import/memo-parser' {
   declare module.exports: any;
 }
 
@@ -275,7 +275,9 @@ declare module 'eslint-plugin-import/lib/core/importType.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/core/importType'>;
 }
 declare module 'eslint-plugin-import/lib/core/staticRequire.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/core/staticRequire'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/core/staticRequire'
+  >;
 }
 declare module 'eslint-plugin-import/lib/docsUrl.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/docsUrl'>;
@@ -284,22 +286,31 @@ declare module 'eslint-plugin-import/lib/ExportMap.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/ExportMap'>;
 }
 declare module 'eslint-plugin-import/lib/importDeclaration.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/importDeclaration'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/importDeclaration'
+  >;
+}
+declare module 'eslint-plugin-import/lib/index' {
+  declare module.exports: $Exports<'eslint-plugin-import/lib'>;
 }
 declare module 'eslint-plugin-import/lib/index.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/index'>;
+  declare module.exports: $Exports<'eslint-plugin-import/lib'>;
 }
 declare module 'eslint-plugin-import/lib/rules/default.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/default'>;
 }
 declare module 'eslint-plugin-import/lib/rules/dynamic-import-chunkname.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/dynamic-import-chunkname'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/dynamic-import-chunkname'
+  >;
 }
 declare module 'eslint-plugin-import/lib/rules/export.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/export'>;
 }
 declare module 'eslint-plugin-import/lib/rules/exports-last.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/exports-last'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/exports-last'
+  >;
 }
 declare module 'eslint-plugin-import/lib/rules/extensions.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/extensions'>;
@@ -308,13 +319,19 @@ declare module 'eslint-plugin-import/lib/rules/first.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/first'>;
 }
 declare module 'eslint-plugin-import/lib/rules/group-exports.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/group-exports'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/group-exports'
+  >;
 }
 declare module 'eslint-plugin-import/lib/rules/imports-first.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/imports-first'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/imports-first'
+  >;
 }
 declare module 'eslint-plugin-import/lib/rules/max-dependencies.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/max-dependencies'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/max-dependencies'
+  >;
 }
 declare module 'eslint-plugin-import/lib/rules/named.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/named'>;
@@ -323,95 +340,152 @@ declare module 'eslint-plugin-import/lib/rules/namespace.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/namespace'>;
 }
 declare module 'eslint-plugin-import/lib/rules/newline-after-import.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/newline-after-import'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/newline-after-import'
+  >;
 }
 declare module 'eslint-plugin-import/lib/rules/no-absolute-path.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-absolute-path'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/no-absolute-path'
+  >;
 }
 declare module 'eslint-plugin-import/lib/rules/no-amd.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-amd'>;
 }
 declare module 'eslint-plugin-import/lib/rules/no-anonymous-default-export.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-anonymous-default-export'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/no-anonymous-default-export'
+  >;
 }
 declare module 'eslint-plugin-import/lib/rules/no-commonjs.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-commonjs'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/no-commonjs'
+  >;
 }
 declare module 'eslint-plugin-import/lib/rules/no-cycle.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-cycle'>;
 }
 declare module 'eslint-plugin-import/lib/rules/no-default-export.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-default-export'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/no-default-export'
+  >;
 }
 declare module 'eslint-plugin-import/lib/rules/no-deprecated.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-deprecated'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/no-deprecated'
+  >;
 }
 declare module 'eslint-plugin-import/lib/rules/no-duplicates.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-duplicates'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/no-duplicates'
+  >;
 }
 declare module 'eslint-plugin-import/lib/rules/no-dynamic-require.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-dynamic-require'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/no-dynamic-require'
+  >;
 }
 declare module 'eslint-plugin-import/lib/rules/no-extraneous-dependencies.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-extraneous-dependencies'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/no-extraneous-dependencies'
+  >;
 }
 declare module 'eslint-plugin-import/lib/rules/no-internal-modules.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-internal-modules'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/no-internal-modules'
+  >;
 }
 declare module 'eslint-plugin-import/lib/rules/no-mutable-exports.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-mutable-exports'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/no-mutable-exports'
+  >;
 }
 declare module 'eslint-plugin-import/lib/rules/no-named-as-default-member.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-named-as-default-member'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/no-named-as-default-member'
+  >;
 }
 declare module 'eslint-plugin-import/lib/rules/no-named-as-default.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-named-as-default'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/no-named-as-default'
+  >;
 }
 declare module 'eslint-plugin-import/lib/rules/no-named-default.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-named-default'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/no-named-default'
+  >;
 }
 declare module 'eslint-plugin-import/lib/rules/no-named-export.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-named-export'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/no-named-export'
+  >;
 }
 declare module 'eslint-plugin-import/lib/rules/no-namespace.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-namespace'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/no-namespace'
+  >;
 }
 declare module 'eslint-plugin-import/lib/rules/no-nodejs-modules.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-nodejs-modules'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/no-nodejs-modules'
+  >;
 }
 declare module 'eslint-plugin-import/lib/rules/no-relative-parent-imports.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-relative-parent-imports'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/no-relative-parent-imports'
+  >;
 }
 declare module 'eslint-plugin-import/lib/rules/no-restricted-paths.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-restricted-paths'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/no-restricted-paths'
+  >;
 }
 declare module 'eslint-plugin-import/lib/rules/no-self-import.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-self-import'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/no-self-import'
+  >;
 }
 declare module 'eslint-plugin-import/lib/rules/no-unassigned-import.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-unassigned-import'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/no-unassigned-import'
+  >;
 }
 declare module 'eslint-plugin-import/lib/rules/no-unresolved.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-unresolved'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/no-unresolved'
+  >;
 }
 declare module 'eslint-plugin-import/lib/rules/no-unused-modules.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-unused-modules'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/no-unused-modules'
+  >;
 }
 declare module 'eslint-plugin-import/lib/rules/no-useless-path-segments.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-useless-path-segments'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/no-useless-path-segments'
+  >;
 }
 declare module 'eslint-plugin-import/lib/rules/no-webpack-loader-syntax.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/no-webpack-loader-syntax'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/no-webpack-loader-syntax'
+  >;
 }
 declare module 'eslint-plugin-import/lib/rules/order.js' {
   declare module.exports: $Exports<'eslint-plugin-import/lib/rules/order'>;
 }
 declare module 'eslint-plugin-import/lib/rules/prefer-default-export.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/prefer-default-export'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/prefer-default-export'
+  >;
 }
 declare module 'eslint-plugin-import/lib/rules/unambiguous.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/lib/rules/unambiguous'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-import/lib/rules/unambiguous'
+  >;
+}
+declare module 'eslint-plugin-import/memo-parser/index' {
+  declare module.exports: $Exports<'eslint-plugin-import/memo-parser'>;
 }
 declare module 'eslint-plugin-import/memo-parser/index.js' {
-  declare module.exports: $Exports<'eslint-plugin-import/memo-parser/index'>;
+  declare module.exports: $Exports<'eslint-plugin-import/memo-parser'>;
 }
