@@ -1,5 +1,5 @@
-// flow-typed signature: ad9f3d2165e9779f8ed708ef640758ea
-// flow-typed version: <<STUB>>/eslint-plugin-jsx-a11y_v6.2.1/flow_v0.99.1
+// flow-typed signature: 750f6e196696e952abecdc8e96cd14ee
+// flow-typed version: <<STUB>>/eslint-plugin-jsx-a11y_v6.2.3/flow_v0.112.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -274,6 +274,10 @@ declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isNonInteractiveRole-t
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isNonLiteralProperty-test' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isSemanticRoleElement-test' {
   declare module.exports: any;
 }
@@ -294,7 +298,7 @@ declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/schemas-test' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jsx-a11y/lib/index' {
+declare module 'eslint-plugin-jsx-a11y/lib' {
   declare module.exports: any;
 }
 
@@ -542,7 +546,7 @@ declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/img' {
   declare module.exports: any;
 }
 
-declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/index' {
+declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles' {
   declare module.exports: any;
 }
 
@@ -650,6 +654,10 @@ declare module 'eslint-plugin-jsx-a11y/lib/util/isNonInteractiveRole' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-jsx-a11y/lib/util/isNonLiteralProperty' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-jsx-a11y/lib/util/isPresentationRole' {
   declare module.exports: any;
 }
@@ -692,503 +700,845 @@ declare module 'eslint-plugin-jsx-a11y/scripts/create-rule' {
 
 // Filename aliases
 declare module 'eslint-plugin-jsx-a11y/__mocks__/genInteractives.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__mocks__/genInteractives'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__mocks__/genInteractives'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__mocks__/IdentifierMock.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__mocks__/IdentifierMock'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__mocks__/IdentifierMock'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__mocks__/JSXAttributeMock.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__mocks__/JSXAttributeMock'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__mocks__/JSXAttributeMock'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__mocks__/JSXElementMock.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__mocks__/JSXElementMock'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__mocks__/JSXElementMock'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__mocks__/JSXExpressionContainerMock.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__mocks__/JSXExpressionContainerMock'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__mocks__/JSXExpressionContainerMock'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__mocks__/JSXSpreadAttributeMock.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__mocks__/JSXSpreadAttributeMock'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__mocks__/JSXSpreadAttributeMock'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__mocks__/JSXTextMock.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__mocks__/JSXTextMock'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__mocks__/JSXTextMock'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__mocks__/LiteralMock.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__mocks__/LiteralMock'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__mocks__/LiteralMock'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/__util__/parserOptionsMapper.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/__util__/parserOptionsMapper'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/__util__/parserOptionsMapper'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/__util__/ruleOptionsMapperFactory.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/__util__/ruleOptionsMapperFactory'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/__util__/ruleOptionsMapperFactory'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/index-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/index-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/index-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/accessible-emoji-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/accessible-emoji-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/accessible-emoji-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/alt-text-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/alt-text-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/alt-text-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/anchor-has-content-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/anchor-has-content-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/anchor-has-content-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/anchor-is-valid-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/anchor-is-valid-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/anchor-is-valid-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-activedescendant-has-tabindex-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-activedescendant-has-tabindex-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-activedescendant-has-tabindex-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-props-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-props-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-props-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-proptypes-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-proptypes-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-proptypes-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-role-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-role-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-role-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-unsupported-elements-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-unsupported-elements-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/aria-unsupported-elements-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/click-events-have-key-events-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/click-events-have-key-events-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/click-events-have-key-events-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/control-has-associated-label-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/control-has-associated-label-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/control-has-associated-label-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/heading-has-content-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/heading-has-content-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/heading-has-content-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/html-has-lang-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/html-has-lang-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/html-has-lang-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/iframe-has-title-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/iframe-has-title-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/iframe-has-title-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/img-redundant-alt-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/img-redundant-alt-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/img-redundant-alt-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/interactive-supports-focus-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/interactive-supports-focus-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/interactive-supports-focus-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/label-has-associated-control-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/label-has-associated-control-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/label-has-associated-control-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/label-has-for-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/label-has-for-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/label-has-for-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/lang-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/lang-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/lang-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/media-has-caption-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/media-has-caption-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/media-has-caption-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/mouse-events-have-key-events-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/mouse-events-have-key-events-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/mouse-events-have-key-events-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/no-access-key-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/no-access-key-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/no-access-key-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/no-autofocus-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/no-autofocus-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/no-autofocus-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/no-distracting-elements-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/no-distracting-elements-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/no-distracting-elements-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/no-interactive-element-to-noninteractive-role-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/no-interactive-element-to-noninteractive-role-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/no-interactive-element-to-noninteractive-role-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/no-noninteractive-element-interactions-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/no-noninteractive-element-interactions-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/no-noninteractive-element-interactions-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/no-noninteractive-element-to-interactive-role-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/no-noninteractive-element-to-interactive-role-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/no-noninteractive-element-to-interactive-role-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/no-noninteractive-tabindex-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/no-noninteractive-tabindex-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/no-noninteractive-tabindex-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/no-onchange-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/no-onchange-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/no-onchange-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/no-redundant-roles-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/no-redundant-roles-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/no-redundant-roles-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/no-static-element-interactions-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/no-static-element-interactions-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/no-static-element-interactions-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/role-has-required-aria-props-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/role-has-required-aria-props-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/role-has-required-aria-props-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/role-supports-aria-props-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/role-supports-aria-props-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/role-supports-aria-props-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/scope-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/scope-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/scope-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/rules/tabindex-no-positive-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/rules/tabindex-no-positive-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/rules/tabindex-no-positive-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/attributesComparator-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/attributesComparator-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/util/attributesComparator-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/getComputedRole-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/getComputedRole-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/util/getComputedRole-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/getExplicitRole-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/getExplicitRole-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/util/getExplicitRole-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/getImplicitRole-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/getImplicitRole-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/util/getImplicitRole-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/getSuggestion-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/getSuggestion-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/util/getSuggestion-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/getTabIndex-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/getTabIndex-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/util/getTabIndex-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/hasAccessibleChild-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/hasAccessibleChild-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/util/hasAccessibleChild-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/implicitRoles/input-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/implicitRoles/input-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/util/implicitRoles/input-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/implicitRoles/menu-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/implicitRoles/menu-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/util/implicitRoles/menu-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/implicitRoles/menuitem-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/implicitRoles/menuitem-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/util/implicitRoles/menuitem-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isAbstractRole-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/isAbstractRole-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/util/isAbstractRole-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isDisabledElement-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/isDisabledElement-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/util/isDisabledElement-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isDOMElement-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/isDOMElement-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/util/isDOMElement-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isInteractiveElement-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/isInteractiveElement-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/util/isInteractiveElement-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isInteractiveRole-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/isInteractiveRole-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/util/isInteractiveRole-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isNonInteractiveElement-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/isNonInteractiveElement-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/util/isNonInteractiveElement-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isNonInteractiveRole-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/isNonInteractiveRole-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/util/isNonInteractiveRole-test'
+  >;
+}
+declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isNonLiteralProperty-test.js' {
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/util/isNonLiteralProperty-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/isSemanticRoleElement-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/isSemanticRoleElement-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/util/isSemanticRoleElement-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/mayContainChildComponent-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/mayContainChildComponent-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/util/mayContainChildComponent-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/mayHaveAccessibleLabel-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/mayHaveAccessibleLabel-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/util/mayHaveAccessibleLabel-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/parserOptionsMapper-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/parserOptionsMapper-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/util/parserOptionsMapper-test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/__tests__/src/util/schemas-test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/__tests__/src/util/schemas-test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/__tests__/src/util/schemas-test'
+  >;
+}
+declare module 'eslint-plugin-jsx-a11y/lib/index' {
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib'>;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/index.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/index'>;
+  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib'>;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/accessible-emoji.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/accessible-emoji'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/accessible-emoji'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/alt-text.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/alt-text'>;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/anchor-has-content.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/anchor-has-content'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/anchor-has-content'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/anchor-is-valid.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/anchor-is-valid'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/anchor-is-valid'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/aria-activedescendant-has-tabindex.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/aria-activedescendant-has-tabindex'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/aria-activedescendant-has-tabindex'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/aria-props.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/aria-props'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/aria-props'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/aria-proptypes.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/aria-proptypes'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/aria-proptypes'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/aria-role.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/aria-role'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/aria-role'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/aria-unsupported-elements.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/aria-unsupported-elements'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/aria-unsupported-elements'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/click-events-have-key-events.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/click-events-have-key-events'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/click-events-have-key-events'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/control-has-associated-label.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/control-has-associated-label'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/control-has-associated-label'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/heading-has-content.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/heading-has-content'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/heading-has-content'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/html-has-lang.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/html-has-lang'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/html-has-lang'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/iframe-has-title.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/iframe-has-title'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/iframe-has-title'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/img-redundant-alt.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/img-redundant-alt'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/img-redundant-alt'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/interactive-supports-focus.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/interactive-supports-focus'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/interactive-supports-focus'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/label-has-associated-control.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/label-has-associated-control'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/label-has-associated-control'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/label-has-for.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/label-has-for'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/label-has-for'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/lang.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/lang'>;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/media-has-caption.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/media-has-caption'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/media-has-caption'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/mouse-events-have-key-events.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/mouse-events-have-key-events'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/mouse-events-have-key-events'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/no-access-key.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/no-access-key'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/no-access-key'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/no-autofocus.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/no-autofocus'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/no-autofocus'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/no-distracting-elements.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/no-distracting-elements'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/no-distracting-elements'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/no-interactive-element-to-noninteractive-role.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/no-interactive-element-to-noninteractive-role'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/no-interactive-element-to-noninteractive-role'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/no-noninteractive-element-interactions.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/no-noninteractive-element-interactions'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/no-noninteractive-element-interactions'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/no-noninteractive-element-to-interactive-role.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/no-noninteractive-element-to-interactive-role'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/no-noninteractive-element-to-interactive-role'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/no-noninteractive-tabindex.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/no-noninteractive-tabindex'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/no-noninteractive-tabindex'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/no-onchange.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/no-onchange'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/no-onchange'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/no-redundant-roles.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/no-redundant-roles'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/no-redundant-roles'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/no-static-element-interactions.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/no-static-element-interactions'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/no-static-element-interactions'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/role-has-required-aria-props.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/role-has-required-aria-props'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/role-has-required-aria-props'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/role-supports-aria-props.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/role-supports-aria-props'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/role-supports-aria-props'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/scope.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/scope'>;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/rules/tabindex-no-positive.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/rules/tabindex-no-positive'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/rules/tabindex-no-positive'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/attributesComparator.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/attributesComparator'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/attributesComparator'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/getComputedRole.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/getComputedRole'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/getComputedRole'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/getExplicitRole.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/getExplicitRole'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/getExplicitRole'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/getImplicitRole.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/getImplicitRole'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/getImplicitRole'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/getSuggestion.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/getSuggestion'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/getSuggestion'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/getTabIndex.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/getTabIndex'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/getTabIndex'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/hasAccessibleChild.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/hasAccessibleChild'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/hasAccessibleChild'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/a.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/a'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/a'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/area.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/area'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/area'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/article.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/article'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/article'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/aside.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/aside'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/aside'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/body.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/body'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/body'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/button.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/button'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/button'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/datalist.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/datalist'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/datalist'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/details.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/details'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/details'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/dialog.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/dialog'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/dialog'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/dl.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/dl'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/dl'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/form.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/form'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/form'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/h1.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/h1'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/h1'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/h2.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/h2'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/h2'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/h3.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/h3'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/h3'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/h4.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/h4'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/h4'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/h5.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/h5'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/h5'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/h6.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/h6'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/h6'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/hr.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/hr'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/hr'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/img.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/img'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/img'
+  >;
+}
+declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/index' {
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/index.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/index'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/input.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/input'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/input'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/li.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/li'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/li'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/link.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/link'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/link'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/menu.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/menu'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/menu'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/menuitem.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/menuitem'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/menuitem'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/meter.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/meter'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/meter'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/nav.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/nav'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/nav'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/ol.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/ol'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/ol'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/option.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/option'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/option'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/output.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/output'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/output'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/progress.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/progress'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/progress'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/section.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/section'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/section'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/select.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/select'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/select'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/tbody.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/tbody'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/tbody'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/textarea.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/textarea'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/textarea'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/tfoot.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/tfoot'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/tfoot'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/thead.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/thead'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/thead'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/implicitRoles/ul.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/implicitRoles/ul'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/implicitRoles/ul'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/isAbstractRole.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/isAbstractRole'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/isAbstractRole'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/isDisabledElement.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/isDisabledElement'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/isDisabledElement'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/isDOMElement.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/isDOMElement'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/isDOMElement'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/isHiddenFromScreenReader.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/isHiddenFromScreenReader'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/isHiddenFromScreenReader'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/isInteractiveElement.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/isInteractiveElement'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/isInteractiveElement'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/isInteractiveRole.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/isInteractiveRole'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/isInteractiveRole'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/isNonInteractiveElement.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/isNonInteractiveElement'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/isNonInteractiveElement'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/isNonInteractiveRole.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/isNonInteractiveRole'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/isNonInteractiveRole'
+  >;
+}
+declare module 'eslint-plugin-jsx-a11y/lib/util/isNonLiteralProperty.js' {
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/isNonLiteralProperty'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/isPresentationRole.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/isPresentationRole'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/isPresentationRole'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/isSemanticRoleElement.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/isSemanticRoleElement'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/isSemanticRoleElement'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/mayContainChildComponent.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/mayContainChildComponent'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/mayContainChildComponent'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/mayHaveAccessibleLabel.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/mayHaveAccessibleLabel'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/lib/util/mayHaveAccessibleLabel'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/lib/util/schemas.js' {
   declare module.exports: $Exports<'eslint-plugin-jsx-a11y/lib/util/schemas'>;
 }
 declare module 'eslint-plugin-jsx-a11y/scripts/addRuleToIndex.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/scripts/addRuleToIndex'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/scripts/addRuleToIndex'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/scripts/boilerplate/doc.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/scripts/boilerplate/doc'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/scripts/boilerplate/doc'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/scripts/boilerplate/rule.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/scripts/boilerplate/rule'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/scripts/boilerplate/rule'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/scripts/boilerplate/test.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/scripts/boilerplate/test'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/scripts/boilerplate/test'
+  >;
 }
 declare module 'eslint-plugin-jsx-a11y/scripts/create-rule.js' {
-  declare module.exports: $Exports<'eslint-plugin-jsx-a11y/scripts/create-rule'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-jsx-a11y/scripts/create-rule'
+  >;
 }
