@@ -1,5 +1,5 @@
-// flow-typed signature: e47e9f69ffe2b6269fe7f99ddb328d9b
-// flow-typed version: <<STUB>>/eslint-plugin-prettier_v3.1.0/flow_v0.99.1
+// flow-typed signature: de62b7eb25cf64aa47bad2c8e638e79e
+// flow-typed version: <<STUB>>/eslint-plugin-prettier_v3.1.1/flow_v0.112.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -28,5 +28,7 @@ declare module 'eslint-plugin-prettier/eslint-plugin-prettier' {
 
 // Filename aliases
 declare module 'eslint-plugin-prettier/eslint-plugin-prettier.js' {
-  declare module.exports: $Exports<'eslint-plugin-prettier/eslint-plugin-prettier'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-prettier/eslint-plugin-prettier'
+  >;
 }
