@@ -1,5 +1,5 @@
-// flow-typed signature: 22826ce310e6f55d96a382962e30fdc5
-// flow-typed version: <<STUB>>/eslint-plugin-promise_v4.1.1/flow_v0.99.1
+// flow-typed signature: 7404e4a36ecf526e40cf0e1951185905
+// flow-typed version: <<STUB>>/eslint-plugin-promise_v4.2.1/flow_v0.112.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -124,34 +124,54 @@ declare module 'eslint-plugin-promise/rules/avoid-new.js' {
   declare module.exports: $Exports<'eslint-plugin-promise/rules/avoid-new'>;
 }
 declare module 'eslint-plugin-promise/rules/catch-or-return.js' {
-  declare module.exports: $Exports<'eslint-plugin-promise/rules/catch-or-return'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-promise/rules/catch-or-return'
+  >;
 }
 declare module 'eslint-plugin-promise/rules/lib/get-docs-url.js' {
-  declare module.exports: $Exports<'eslint-plugin-promise/rules/lib/get-docs-url'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-promise/rules/lib/get-docs-url'
+  >;
 }
 declare module 'eslint-plugin-promise/rules/lib/has-promise-callback.js' {
-  declare module.exports: $Exports<'eslint-plugin-promise/rules/lib/has-promise-callback'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-promise/rules/lib/has-promise-callback'
+  >;
 }
 declare module 'eslint-plugin-promise/rules/lib/is-callback.js' {
-  declare module.exports: $Exports<'eslint-plugin-promise/rules/lib/is-callback'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-promise/rules/lib/is-callback'
+  >;
 }
 declare module 'eslint-plugin-promise/rules/lib/is-inside-callback.js' {
-  declare module.exports: $Exports<'eslint-plugin-promise/rules/lib/is-inside-callback'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-promise/rules/lib/is-inside-callback'
+  >;
 }
 declare module 'eslint-plugin-promise/rules/lib/is-inside-promise.js' {
-  declare module.exports: $Exports<'eslint-plugin-promise/rules/lib/is-inside-promise'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-promise/rules/lib/is-inside-promise'
+  >;
 }
 declare module 'eslint-plugin-promise/rules/lib/is-named-callback.js' {
-  declare module.exports: $Exports<'eslint-plugin-promise/rules/lib/is-named-callback'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-promise/rules/lib/is-named-callback'
+  >;
 }
 declare module 'eslint-plugin-promise/rules/lib/is-promise.js' {
-  declare module.exports: $Exports<'eslint-plugin-promise/rules/lib/is-promise'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-promise/rules/lib/is-promise'
+  >;
 }
 declare module 'eslint-plugin-promise/rules/lib/promise-statics.js' {
-  declare module.exports: $Exports<'eslint-plugin-promise/rules/lib/promise-statics'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-promise/rules/lib/promise-statics'
+  >;
 }
 declare module 'eslint-plugin-promise/rules/no-callback-in-promise.js' {
-  declare module.exports: $Exports<'eslint-plugin-promise/rules/no-callback-in-promise'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-promise/rules/no-callback-in-promise'
+  >;
 }
 declare module 'eslint-plugin-promise/rules/no-native.js' {
   declare module.exports: $Exports<'eslint-plugin-promise/rules/no-native'>;
@@ -160,25 +180,37 @@ declare module 'eslint-plugin-promise/rules/no-nesting.js' {
   declare module.exports: $Exports<'eslint-plugin-promise/rules/no-nesting'>;
 }
 declare module 'eslint-plugin-promise/rules/no-new-statics.js' {
-  declare module.exports: $Exports<'eslint-plugin-promise/rules/no-new-statics'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-promise/rules/no-new-statics'
+  >;
 }
 declare module 'eslint-plugin-promise/rules/no-promise-in-callback.js' {
-  declare module.exports: $Exports<'eslint-plugin-promise/rules/no-promise-in-callback'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-promise/rules/no-promise-in-callback'
+  >;
 }
 declare module 'eslint-plugin-promise/rules/no-return-in-finally.js' {
-  declare module.exports: $Exports<'eslint-plugin-promise/rules/no-return-in-finally'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-promise/rules/no-return-in-finally'
+  >;
 }
 declare module 'eslint-plugin-promise/rules/no-return-wrap.js' {
-  declare module.exports: $Exports<'eslint-plugin-promise/rules/no-return-wrap'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-promise/rules/no-return-wrap'
+  >;
 }
 declare module 'eslint-plugin-promise/rules/param-names.js' {
   declare module.exports: $Exports<'eslint-plugin-promise/rules/param-names'>;
 }
 declare module 'eslint-plugin-promise/rules/prefer-await-to-callbacks.js' {
-  declare module.exports: $Exports<'eslint-plugin-promise/rules/prefer-await-to-callbacks'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-promise/rules/prefer-await-to-callbacks'
+  >;
 }
 declare module 'eslint-plugin-promise/rules/prefer-await-to-then.js' {
-  declare module.exports: $Exports<'eslint-plugin-promise/rules/prefer-await-to-then'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-promise/rules/prefer-await-to-then'
+  >;
 }
 declare module 'eslint-plugin-promise/rules/valid-params.js' {
   declare module.exports: $Exports<'eslint-plugin-promise/rules/valid-params'>;
