@@ -1,5 +1,5 @@
-// flow-typed signature: 07a97d00cc1ecc3b3c57bc329c9a6af2
-// flow-typed version: <<STUB>>/eslint-plugin-react_v7.13.0/flow_v0.99.1
+// flow-typed signature: 2fe58e53f903d0bbe77c95a360fba4c2
+// flow-typed version: <<STUB>>/eslint-plugin-react_v7.16.0/flow_v0.112.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -82,6 +82,10 @@ declare module 'eslint-plugin-react/lib/rules/jsx-curly-brace-presence' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-react/lib/rules/jsx-curly-newline' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-react/lib/rules/jsx-curly-spacing' {
   declare module.exports: any;
 }
@@ -147,6 +151,10 @@ declare module 'eslint-plugin-react/lib/rules/jsx-no-target-blank' {
 }
 
 declare module 'eslint-plugin-react/lib/rules/jsx-no-undef' {
+  declare module.exports: any;
+}
+
+declare module 'eslint-plugin-react/lib/rules/jsx-no-useless-fragment' {
   declare module.exports: any;
 }
 
@@ -406,6 +414,10 @@ declare module 'eslint-plugin-react/lib/util/propTypes' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-react/lib/util/propTypesSort' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-react/lib/util/propWrapper' {
   declare module.exports: any;
 }
@@ -430,70 +442,119 @@ declare module 'eslint-plugin-react/index.js' {
   declare module.exports: $Exports<'eslint-plugin-react'>;
 }
 declare module 'eslint-plugin-react/lib/rules/boolean-prop-naming.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/boolean-prop-naming'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/boolean-prop-naming'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/button-has-type.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/button-has-type'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/button-has-type'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/default-props-match-prop-types.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/default-props-match-prop-types'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/default-props-match-prop-types'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/destructuring-assignment.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/destructuring-assignment'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/destructuring-assignment'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/display-name.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/display-name'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/display-name'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/forbid-component-props.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/forbid-component-props'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/forbid-component-props'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/forbid-dom-props.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/forbid-dom-props'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/forbid-dom-props'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/forbid-elements.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/forbid-elements'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/forbid-elements'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/forbid-foreign-prop-types.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/forbid-foreign-prop-types'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/forbid-foreign-prop-types'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/forbid-prop-types.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/forbid-prop-types'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/forbid-prop-types'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-boolean-value.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-boolean-value'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-boolean-value'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-child-element-spacing.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-child-element-spacing'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-child-element-spacing'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-closing-bracket-location.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-closing-bracket-location'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-closing-bracket-location'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-closing-tag-location.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-closing-tag-location'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-closing-tag-location'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-curly-brace-presence.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-curly-brace-presence'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-curly-brace-presence'
+  >;
+}
+declare module 'eslint-plugin-react/lib/rules/jsx-curly-newline.js' {
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-curly-newline'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-curly-spacing.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-curly-spacing'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-curly-spacing'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-equals-spacing.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-equals-spacing'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-equals-spacing'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-filename-extension.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-filename-extension'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-filename-extension'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-first-prop-new-line.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-first-prop-new-line'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-first-prop-new-line'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-fragments.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-fragments'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-fragments'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-handler-names.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-handler-names'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-handler-names'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-indent-props.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-indent-props'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-indent-props'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-indent.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-indent'>;
@@ -502,178 +563,287 @@ declare module 'eslint-plugin-react/lib/rules/jsx-key.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-key'>;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-max-depth.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-max-depth'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-max-depth'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-max-props-per-line.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-max-props-per-line'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-max-props-per-line'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-no-bind.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-no-bind'>;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-no-comment-textnodes.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-no-comment-textnodes'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-no-comment-textnodes'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-no-duplicate-props.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-no-duplicate-props'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-no-duplicate-props'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-no-literals.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-no-literals'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-no-literals'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-no-target-blank.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-no-target-blank'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-no-target-blank'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-no-undef.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-no-undef'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-no-undef'
+  >;
+}
+declare module 'eslint-plugin-react/lib/rules/jsx-no-useless-fragment.js' {
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-no-useless-fragment'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-one-expression-per-line.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-one-expression-per-line'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-one-expression-per-line'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-pascal-case.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-pascal-case'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-pascal-case'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-props-no-multi-spaces.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-props-no-multi-spaces'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-props-no-multi-spaces'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-props-no-spreading.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-props-no-spreading'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-props-no-spreading'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-sort-default-props.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-sort-default-props'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-sort-default-props'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-sort-props.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-sort-props'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-sort-props'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-space-before-closing.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-space-before-closing'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-space-before-closing'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-tag-spacing.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-tag-spacing'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-tag-spacing'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-uses-react.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-uses-react'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-uses-react'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-uses-vars.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-uses-vars'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-uses-vars'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/jsx-wrap-multilines.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/jsx-wrap-multilines'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/jsx-wrap-multilines'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/no-access-state-in-setstate.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-access-state-in-setstate'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/no-access-state-in-setstate'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/no-array-index-key.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-array-index-key'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/no-array-index-key'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/no-children-prop.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-children-prop'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/no-children-prop'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/no-danger-with-children.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-danger-with-children'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/no-danger-with-children'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/no-danger.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-danger'>;
 }
 declare module 'eslint-plugin-react/lib/rules/no-deprecated.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-deprecated'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/no-deprecated'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/no-did-mount-set-state.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-did-mount-set-state'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/no-did-mount-set-state'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/no-did-update-set-state.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-did-update-set-state'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/no-did-update-set-state'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/no-direct-mutation-state.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-direct-mutation-state'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/no-direct-mutation-state'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/no-find-dom-node.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-find-dom-node'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/no-find-dom-node'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/no-is-mounted.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-is-mounted'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/no-is-mounted'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/no-multi-comp.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-multi-comp'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/no-multi-comp'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/no-redundant-should-component-update.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-redundant-should-component-update'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/no-redundant-should-component-update'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/no-render-return-value.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-render-return-value'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/no-render-return-value'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/no-set-state.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-set-state'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/no-set-state'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/no-string-refs.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-string-refs'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/no-string-refs'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/no-this-in-sfc.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-this-in-sfc'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/no-this-in-sfc'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/no-typos.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-typos'>;
 }
 declare module 'eslint-plugin-react/lib/rules/no-unescaped-entities.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-unescaped-entities'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/no-unescaped-entities'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/no-unknown-property.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-unknown-property'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/no-unknown-property'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/no-unsafe.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-unsafe'>;
 }
 declare module 'eslint-plugin-react/lib/rules/no-unused-prop-types.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-unused-prop-types'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/no-unused-prop-types'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/no-unused-state.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-unused-state'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/no-unused-state'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/no-will-update-set-state.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/no-will-update-set-state'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/no-will-update-set-state'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/prefer-es6-class.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/prefer-es6-class'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/prefer-es6-class'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/prefer-read-only-props.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/prefer-read-only-props'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/prefer-read-only-props'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/prefer-stateless-function.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/prefer-stateless-function'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/prefer-stateless-function'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/prop-types.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/prop-types'>;
 }
 declare module 'eslint-plugin-react/lib/rules/react-in-jsx-scope.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/react-in-jsx-scope'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/react-in-jsx-scope'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/require-default-props.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/require-default-props'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/require-default-props'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/require-optimization.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/require-optimization'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/require-optimization'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/require-render-return.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/require-render-return'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/require-render-return'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/self-closing-comp.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/self-closing-comp'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/self-closing-comp'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/sort-comp.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/rules/sort-comp'>;
 }
 declare module 'eslint-plugin-react/lib/rules/sort-prop-types.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/sort-prop-types'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/sort-prop-types'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/state-in-constructor.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/state-in-constructor'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/state-in-constructor'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/static-property-placement.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/static-property-placement'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/static-property-placement'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/style-prop-object.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/style-prop-object'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/style-prop-object'
+  >;
 }
 declare module 'eslint-plugin-react/lib/rules/void-dom-elements-no-children.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/rules/void-dom-elements-no-children'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/rules/void-dom-elements-no-children'
+  >;
 }
 declare module 'eslint-plugin-react/lib/util/annotations.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/util/annotations'>;
@@ -694,19 +864,25 @@ declare module 'eslint-plugin-react/lib/util/error.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/util/error'>;
 }
 declare module 'eslint-plugin-react/lib/util/getTokenBeforeClosingBracket.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/util/getTokenBeforeClosingBracket'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/util/getTokenBeforeClosingBracket'
+  >;
 }
 declare module 'eslint-plugin-react/lib/util/jsx.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/util/jsx'>;
 }
 declare module 'eslint-plugin-react/lib/util/linkComponents.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/util/linkComponents'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/util/linkComponents'
+  >;
 }
 declare module 'eslint-plugin-react/lib/util/log.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/util/log'>;
 }
 declare module 'eslint-plugin-react/lib/util/makeNoMethodSetStateRule.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/util/makeNoMethodSetStateRule'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/util/makeNoMethodSetStateRule'
+  >;
 }
 declare module 'eslint-plugin-react/lib/util/pragma.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/util/pragma'>;
@@ -717,11 +893,18 @@ declare module 'eslint-plugin-react/lib/util/props.js' {
 declare module 'eslint-plugin-react/lib/util/propTypes.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/util/propTypes'>;
 }
+declare module 'eslint-plugin-react/lib/util/propTypesSort.js' {
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/util/propTypesSort'
+  >;
+}
 declare module 'eslint-plugin-react/lib/util/propWrapper.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/util/propWrapper'>;
 }
 declare module 'eslint-plugin-react/lib/util/usedPropTypes.js' {
-  declare module.exports: $Exports<'eslint-plugin-react/lib/util/usedPropTypes'>;
+  declare module.exports: $Exports<
+    'eslint-plugin-react/lib/util/usedPropTypes'
+  >;
 }
 declare module 'eslint-plugin-react/lib/util/variable.js' {
   declare module.exports: $Exports<'eslint-plugin-react/lib/util/variable'>;
