@@ -1,5 +1,5 @@
-// flow-typed signature: 44224b3f5a70ccd8a85fda0b7f5cf4c8
-// flow-typed version: <<STUB>>/eslint_v^5.16.0/flow_v0.99.1
+// flow-typed signature: 27d48a13993f848712a7fd8b4cbe0399
+// flow-typed version: <<STUB>>/eslint_v^6.6.0/flow_v0.112.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -50,7 +50,99 @@ declare module 'eslint/lib/api' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/built-in-rules-index' {
+declare module 'eslint/lib/cli-engine/cascading-config-array-factory' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/cli-engine' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/config-array-factory' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/config-array/config-array' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/config-array/config-dependency' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/config-array/extracted-config' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/config-array' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/config-array/override-tester' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/file-enumerator' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/formatters/checkstyle' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/formatters/codeframe' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/formatters/compact' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/formatters/html' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/formatters/jslint-xml' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/formatters/json-with-metadata' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/formatters/json' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/formatters/junit' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/formatters/stylish' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/formatters/table' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/formatters/tap' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/formatters/unix' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/formatters/visualstudio' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/hash' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/ignored-paths' {
   declare module.exports: any;
 }
 
@@ -58,127 +150,79 @@ declare module 'eslint/lib/cli-engine' {
   declare module.exports: any;
 }
 
+declare module 'eslint/lib/cli-engine/lint-result-cache' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/load-rules' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/cli-engine/xml-escape' {
+  declare module.exports: any;
+}
+
 declare module 'eslint/lib/cli' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/code-path-analysis/code-path-analyzer' {
+declare module 'eslint/lib/init/autoconfig' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/code-path-analysis/code-path-segment' {
+declare module 'eslint/lib/init/config-file' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/code-path-analysis/code-path-state' {
+declare module 'eslint/lib/init/config-initializer' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/code-path-analysis/code-path' {
+declare module 'eslint/lib/init/config-rule' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/code-path-analysis/debug-helpers' {
+declare module 'eslint/lib/init/npm-utils' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/code-path-analysis/fork-context' {
+declare module 'eslint/lib/init/source-code-utils' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/code-path-analysis/id-generator' {
+declare module 'eslint/lib/linter/apply-disable-directives' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/config' {
+declare module 'eslint/lib/linter/code-path-analysis/code-path-analyzer' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/config/autoconfig' {
+declare module 'eslint/lib/linter/code-path-analysis/code-path-segment' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/config/config-cache' {
+declare module 'eslint/lib/linter/code-path-analysis/code-path-state' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/config/config-file' {
+declare module 'eslint/lib/linter/code-path-analysis/code-path' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/config/config-initializer' {
+declare module 'eslint/lib/linter/code-path-analysis/debug-helpers' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/config/config-ops' {
+declare module 'eslint/lib/linter/code-path-analysis/fork-context' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/config/config-rule' {
+declare module 'eslint/lib/linter/code-path-analysis/id-generator' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/config/config-validator' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/config/environments' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/config/plugins' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/formatters/checkstyle' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/formatters/codeframe' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/formatters/compact' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/formatters/html' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/formatters/jslint-xml' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/formatters/json-with-metadata' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/formatters/json' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/formatters/junit' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/formatters/stylish' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/formatters/table' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/formatters/tap' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/formatters/unix' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/formatters/visualstudio' {
+declare module 'eslint/lib/linter/config-comment-parser' {
   declare module.exports: any;
 }
 
@@ -186,7 +230,39 @@ declare module 'eslint/lib/linter' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/load-rules' {
+declare module 'eslint/lib/linter/interpolate' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/linter/linter' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/linter/node-event-generator' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/linter/report-translator' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/linter/rule-fixer' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/linter/rules' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/linter/safe-emitter' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/linter/source-code-fixer' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/linter/timing' {
   declare module.exports: any;
 }
 
@@ -194,7 +270,11 @@ declare module 'eslint/lib/options' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/rules' {
+declare module 'eslint/lib/rule-tester' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rule-tester/rule-tester' {
   declare module.exports: any;
 }
 
@@ -298,6 +378,10 @@ declare module 'eslint/lib/rules/default-case' {
   declare module.exports: any;
 }
 
+declare module 'eslint/lib/rules/default-param-last' {
+  declare module.exports: any;
+}
+
 declare module 'eslint/lib/rules/dot-location' {
   declare module.exports: any;
 }
@@ -331,6 +415,10 @@ declare module 'eslint/lib/rules/func-names' {
 }
 
 declare module 'eslint/lib/rules/func-style' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/function-call-argument-newline' {
   declare module.exports: any;
 }
 
@@ -379,6 +467,10 @@ declare module 'eslint/lib/rules/indent-legacy' {
 }
 
 declare module 'eslint/lib/rules/indent' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules' {
   declare module.exports: any;
 }
 
@@ -667,6 +759,10 @@ declare module 'eslint/lib/rules/no-implicit-globals' {
 }
 
 declare module 'eslint/lib/rules/no-implied-eval' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/no-import-assign' {
   declare module.exports: any;
 }
 
@@ -1110,6 +1206,10 @@ declare module 'eslint/lib/rules/prefer-reflect' {
   declare module.exports: any;
 }
 
+declare module 'eslint/lib/rules/prefer-regex-literals' {
+  declare module.exports: any;
+}
+
 declare module 'eslint/lib/rules/prefer-rest-params' {
   declare module.exports: any;
 }
@@ -1234,6 +1334,46 @@ declare module 'eslint/lib/rules/use-isnan' {
   declare module.exports: any;
 }
 
+declare module 'eslint/lib/rules/utils/ast-utils' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/utils/fix-tracker' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/utils/keywords' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/utils/lazy-loading-rule-map' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/utils/patterns/letters' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/utils/unicode' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/utils/unicode/is-combining-character' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/utils/unicode/is-emoji-modifier' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/utils/unicode/is-regional-indicator-symbol' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/utils/unicode/is-surrogate-pair' {
+  declare module.exports: any;
+}
+
 declare module 'eslint/lib/rules/valid-jsdoc' {
   declare module.exports: any;
 }
@@ -1262,195 +1402,103 @@ declare module 'eslint/lib/rules/yoda' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/testers/rule-tester' {
+declare module 'eslint/lib/shared/ajv' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/token-store/backward-token-comment-cursor' {
+declare module 'eslint/lib/shared/ast-utils' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/token-store/backward-token-cursor' {
+declare module 'eslint/lib/shared/config-ops' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/token-store/cursor' {
+declare module 'eslint/lib/shared/config-validator' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/token-store/cursors' {
+declare module 'eslint/lib/shared/logging' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/token-store/decorative-cursor' {
+declare module 'eslint/lib/shared/naming' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/token-store/filter-cursor' {
+declare module 'eslint/lib/shared/relative-module-resolver' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/token-store/forward-token-comment-cursor' {
+declare module 'eslint/lib/shared/runtime-info' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/token-store/forward-token-cursor' {
+declare module 'eslint/lib/shared/traverser' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/token-store/index' {
+declare module 'eslint/lib/shared/types' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/token-store/limit-cursor' {
+declare module 'eslint/lib/source-code' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/token-store/padded-token-cursor' {
+declare module 'eslint/lib/source-code/source-code' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/token-store/skip-cursor' {
+declare module 'eslint/lib/source-code/token-store/backward-token-comment-cursor' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/token-store/utils' {
+declare module 'eslint/lib/source-code/token-store/backward-token-cursor' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/ajv' {
+declare module 'eslint/lib/source-code/token-store/cursor' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/apply-disable-directives' {
+declare module 'eslint/lib/source-code/token-store/cursors' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/ast-utils' {
+declare module 'eslint/lib/source-code/token-store/decorative-cursor' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/config-comment-parser' {
+declare module 'eslint/lib/source-code/token-store/filter-cursor' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/file-finder' {
+declare module 'eslint/lib/source-code/token-store/forward-token-comment-cursor' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/fix-tracker' {
+declare module 'eslint/lib/source-code/token-store/forward-token-cursor' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/glob-utils' {
+declare module 'eslint/lib/source-code/token-store' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/glob' {
+declare module 'eslint/lib/source-code/token-store/limit-cursor' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/hash' {
+declare module 'eslint/lib/source-code/token-store/padded-token-cursor' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/ignored-paths' {
+declare module 'eslint/lib/source-code/token-store/skip-cursor' {
   declare module.exports: any;
 }
 
-declare module 'eslint/lib/util/interpolate' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/util/keywords' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/util/lint-result-cache' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/util/logging' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/util/module-resolver' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/util/naming' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/util/node-event-generator' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/util/npm-utils' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/util/path-utils' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/util/patterns/letters' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/util/report-translator' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/util/rule-fixer' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/util/safe-emitter' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/util/source-code-fixer' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/util/source-code-utils' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/util/source-code' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/util/timing' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/util/traverser' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/util/unicode/index' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/util/unicode/is-combining-character' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/util/unicode/is-emoji-modifier' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/util/unicode/is-regional-indicator-symbol' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/util/unicode/is-surrogate-pair' {
-  declare module.exports: any;
-}
-
-declare module 'eslint/lib/util/xml-escape' {
+declare module 'eslint/lib/source-code/token-store/utils' {
   declare module.exports: any;
 }
 
@@ -1476,116 +1524,226 @@ declare module 'eslint/conf/eslint-recommended.js' {
 declare module 'eslint/lib/api.js' {
   declare module.exports: $Exports<'eslint/lib/api'>;
 }
-declare module 'eslint/lib/built-in-rules-index.js' {
-  declare module.exports: $Exports<'eslint/lib/built-in-rules-index'>;
+declare module 'eslint/lib/cli-engine/cascading-config-array-factory.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/cli-engine/cascading-config-array-factory'
+  >;
 }
-declare module 'eslint/lib/cli-engine.js' {
+declare module 'eslint/lib/cli-engine/cli-engine.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/cli-engine'>;
+}
+declare module 'eslint/lib/cli-engine/config-array-factory.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/cli-engine/config-array-factory'
+  >;
+}
+declare module 'eslint/lib/cli-engine/config-array/config-array.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/cli-engine/config-array/config-array'
+  >;
+}
+declare module 'eslint/lib/cli-engine/config-array/config-dependency.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/cli-engine/config-array/config-dependency'
+  >;
+}
+declare module 'eslint/lib/cli-engine/config-array/extracted-config.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/cli-engine/config-array/extracted-config'
+  >;
+}
+declare module 'eslint/lib/cli-engine/config-array/index' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/config-array'>;
+}
+declare module 'eslint/lib/cli-engine/config-array/index.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/config-array'>;
+}
+declare module 'eslint/lib/cli-engine/config-array/override-tester.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/cli-engine/config-array/override-tester'
+  >;
+}
+declare module 'eslint/lib/cli-engine/file-enumerator.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/file-enumerator'>;
+}
+declare module 'eslint/lib/cli-engine/formatters/checkstyle.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/cli-engine/formatters/checkstyle'
+  >;
+}
+declare module 'eslint/lib/cli-engine/formatters/codeframe.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/cli-engine/formatters/codeframe'
+  >;
+}
+declare module 'eslint/lib/cli-engine/formatters/compact.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/formatters/compact'>;
+}
+declare module 'eslint/lib/cli-engine/formatters/html.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/formatters/html'>;
+}
+declare module 'eslint/lib/cli-engine/formatters/jslint-xml.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/cli-engine/formatters/jslint-xml'
+  >;
+}
+declare module 'eslint/lib/cli-engine/formatters/json-with-metadata.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/cli-engine/formatters/json-with-metadata'
+  >;
+}
+declare module 'eslint/lib/cli-engine/formatters/json.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/formatters/json'>;
+}
+declare module 'eslint/lib/cli-engine/formatters/junit.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/formatters/junit'>;
+}
+declare module 'eslint/lib/cli-engine/formatters/stylish.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/formatters/stylish'>;
+}
+declare module 'eslint/lib/cli-engine/formatters/table.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/formatters/table'>;
+}
+declare module 'eslint/lib/cli-engine/formatters/tap.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/formatters/tap'>;
+}
+declare module 'eslint/lib/cli-engine/formatters/unix.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/formatters/unix'>;
+}
+declare module 'eslint/lib/cli-engine/formatters/visualstudio.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/cli-engine/formatters/visualstudio'
+  >;
+}
+declare module 'eslint/lib/cli-engine/hash.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/hash'>;
+}
+declare module 'eslint/lib/cli-engine/ignored-paths.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/ignored-paths'>;
+}
+declare module 'eslint/lib/cli-engine/index' {
   declare module.exports: $Exports<'eslint/lib/cli-engine'>;
+}
+declare module 'eslint/lib/cli-engine/index.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine'>;
+}
+declare module 'eslint/lib/cli-engine/lint-result-cache.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/lint-result-cache'>;
+}
+declare module 'eslint/lib/cli-engine/load-rules.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/load-rules'>;
+}
+declare module 'eslint/lib/cli-engine/xml-escape.js' {
+  declare module.exports: $Exports<'eslint/lib/cli-engine/xml-escape'>;
 }
 declare module 'eslint/lib/cli.js' {
   declare module.exports: $Exports<'eslint/lib/cli'>;
 }
-declare module 'eslint/lib/code-path-analysis/code-path-analyzer.js' {
-  declare module.exports: $Exports<'eslint/lib/code-path-analysis/code-path-analyzer'>;
+declare module 'eslint/lib/init/autoconfig.js' {
+  declare module.exports: $Exports<'eslint/lib/init/autoconfig'>;
 }
-declare module 'eslint/lib/code-path-analysis/code-path-segment.js' {
-  declare module.exports: $Exports<'eslint/lib/code-path-analysis/code-path-segment'>;
+declare module 'eslint/lib/init/config-file.js' {
+  declare module.exports: $Exports<'eslint/lib/init/config-file'>;
 }
-declare module 'eslint/lib/code-path-analysis/code-path-state.js' {
-  declare module.exports: $Exports<'eslint/lib/code-path-analysis/code-path-state'>;
+declare module 'eslint/lib/init/config-initializer.js' {
+  declare module.exports: $Exports<'eslint/lib/init/config-initializer'>;
 }
-declare module 'eslint/lib/code-path-analysis/code-path.js' {
-  declare module.exports: $Exports<'eslint/lib/code-path-analysis/code-path'>;
+declare module 'eslint/lib/init/config-rule.js' {
+  declare module.exports: $Exports<'eslint/lib/init/config-rule'>;
 }
-declare module 'eslint/lib/code-path-analysis/debug-helpers.js' {
-  declare module.exports: $Exports<'eslint/lib/code-path-analysis/debug-helpers'>;
+declare module 'eslint/lib/init/npm-utils.js' {
+  declare module.exports: $Exports<'eslint/lib/init/npm-utils'>;
 }
-declare module 'eslint/lib/code-path-analysis/fork-context.js' {
-  declare module.exports: $Exports<'eslint/lib/code-path-analysis/fork-context'>;
+declare module 'eslint/lib/init/source-code-utils.js' {
+  declare module.exports: $Exports<'eslint/lib/init/source-code-utils'>;
 }
-declare module 'eslint/lib/code-path-analysis/id-generator.js' {
-  declare module.exports: $Exports<'eslint/lib/code-path-analysis/id-generator'>;
+declare module 'eslint/lib/linter/apply-disable-directives.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/linter/apply-disable-directives'
+  >;
 }
-declare module 'eslint/lib/config.js' {
-  declare module.exports: $Exports<'eslint/lib/config'>;
+declare module 'eslint/lib/linter/code-path-analysis/code-path-analyzer.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/linter/code-path-analysis/code-path-analyzer'
+  >;
 }
-declare module 'eslint/lib/config/autoconfig.js' {
-  declare module.exports: $Exports<'eslint/lib/config/autoconfig'>;
+declare module 'eslint/lib/linter/code-path-analysis/code-path-segment.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/linter/code-path-analysis/code-path-segment'
+  >;
 }
-declare module 'eslint/lib/config/config-cache.js' {
-  declare module.exports: $Exports<'eslint/lib/config/config-cache'>;
+declare module 'eslint/lib/linter/code-path-analysis/code-path-state.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/linter/code-path-analysis/code-path-state'
+  >;
 }
-declare module 'eslint/lib/config/config-file.js' {
-  declare module.exports: $Exports<'eslint/lib/config/config-file'>;
+declare module 'eslint/lib/linter/code-path-analysis/code-path.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/linter/code-path-analysis/code-path'
+  >;
 }
-declare module 'eslint/lib/config/config-initializer.js' {
-  declare module.exports: $Exports<'eslint/lib/config/config-initializer'>;
+declare module 'eslint/lib/linter/code-path-analysis/debug-helpers.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/linter/code-path-analysis/debug-helpers'
+  >;
 }
-declare module 'eslint/lib/config/config-ops.js' {
-  declare module.exports: $Exports<'eslint/lib/config/config-ops'>;
+declare module 'eslint/lib/linter/code-path-analysis/fork-context.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/linter/code-path-analysis/fork-context'
+  >;
 }
-declare module 'eslint/lib/config/config-rule.js' {
-  declare module.exports: $Exports<'eslint/lib/config/config-rule'>;
+declare module 'eslint/lib/linter/code-path-analysis/id-generator.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/linter/code-path-analysis/id-generator'
+  >;
 }
-declare module 'eslint/lib/config/config-validator.js' {
-  declare module.exports: $Exports<'eslint/lib/config/config-validator'>;
+declare module 'eslint/lib/linter/config-comment-parser.js' {
+  declare module.exports: $Exports<'eslint/lib/linter/config-comment-parser'>;
 }
-declare module 'eslint/lib/config/environments.js' {
-  declare module.exports: $Exports<'eslint/lib/config/environments'>;
-}
-declare module 'eslint/lib/config/plugins.js' {
-  declare module.exports: $Exports<'eslint/lib/config/plugins'>;
-}
-declare module 'eslint/lib/formatters/checkstyle.js' {
-  declare module.exports: $Exports<'eslint/lib/formatters/checkstyle'>;
-}
-declare module 'eslint/lib/formatters/codeframe.js' {
-  declare module.exports: $Exports<'eslint/lib/formatters/codeframe'>;
-}
-declare module 'eslint/lib/formatters/compact.js' {
-  declare module.exports: $Exports<'eslint/lib/formatters/compact'>;
-}
-declare module 'eslint/lib/formatters/html.js' {
-  declare module.exports: $Exports<'eslint/lib/formatters/html'>;
-}
-declare module 'eslint/lib/formatters/jslint-xml.js' {
-  declare module.exports: $Exports<'eslint/lib/formatters/jslint-xml'>;
-}
-declare module 'eslint/lib/formatters/json-with-metadata.js' {
-  declare module.exports: $Exports<'eslint/lib/formatters/json-with-metadata'>;
-}
-declare module 'eslint/lib/formatters/json.js' {
-  declare module.exports: $Exports<'eslint/lib/formatters/json'>;
-}
-declare module 'eslint/lib/formatters/junit.js' {
-  declare module.exports: $Exports<'eslint/lib/formatters/junit'>;
-}
-declare module 'eslint/lib/formatters/stylish.js' {
-  declare module.exports: $Exports<'eslint/lib/formatters/stylish'>;
-}
-declare module 'eslint/lib/formatters/table.js' {
-  declare module.exports: $Exports<'eslint/lib/formatters/table'>;
-}
-declare module 'eslint/lib/formatters/tap.js' {
-  declare module.exports: $Exports<'eslint/lib/formatters/tap'>;
-}
-declare module 'eslint/lib/formatters/unix.js' {
-  declare module.exports: $Exports<'eslint/lib/formatters/unix'>;
-}
-declare module 'eslint/lib/formatters/visualstudio.js' {
-  declare module.exports: $Exports<'eslint/lib/formatters/visualstudio'>;
-}
-declare module 'eslint/lib/linter.js' {
+declare module 'eslint/lib/linter/index' {
   declare module.exports: $Exports<'eslint/lib/linter'>;
 }
-declare module 'eslint/lib/load-rules.js' {
-  declare module.exports: $Exports<'eslint/lib/load-rules'>;
+declare module 'eslint/lib/linter/index.js' {
+  declare module.exports: $Exports<'eslint/lib/linter'>;
+}
+declare module 'eslint/lib/linter/interpolate.js' {
+  declare module.exports: $Exports<'eslint/lib/linter/interpolate'>;
+}
+declare module 'eslint/lib/linter/linter.js' {
+  declare module.exports: $Exports<'eslint/lib/linter/linter'>;
+}
+declare module 'eslint/lib/linter/node-event-generator.js' {
+  declare module.exports: $Exports<'eslint/lib/linter/node-event-generator'>;
+}
+declare module 'eslint/lib/linter/report-translator.js' {
+  declare module.exports: $Exports<'eslint/lib/linter/report-translator'>;
+}
+declare module 'eslint/lib/linter/rule-fixer.js' {
+  declare module.exports: $Exports<'eslint/lib/linter/rule-fixer'>;
+}
+declare module 'eslint/lib/linter/rules.js' {
+  declare module.exports: $Exports<'eslint/lib/linter/rules'>;
+}
+declare module 'eslint/lib/linter/safe-emitter.js' {
+  declare module.exports: $Exports<'eslint/lib/linter/safe-emitter'>;
+}
+declare module 'eslint/lib/linter/source-code-fixer.js' {
+  declare module.exports: $Exports<'eslint/lib/linter/source-code-fixer'>;
+}
+declare module 'eslint/lib/linter/timing.js' {
+  declare module.exports: $Exports<'eslint/lib/linter/timing'>;
 }
 declare module 'eslint/lib/options.js' {
   declare module.exports: $Exports<'eslint/lib/options'>;
 }
-declare module 'eslint/lib/rules.js' {
-  declare module.exports: $Exports<'eslint/lib/rules'>;
+declare module 'eslint/lib/rule-tester/index' {
+  declare module.exports: $Exports<'eslint/lib/rule-tester'>;
+}
+declare module 'eslint/lib/rule-tester/index.js' {
+  declare module.exports: $Exports<'eslint/lib/rule-tester'>;
+}
+declare module 'eslint/lib/rule-tester/rule-tester.js' {
+  declare module.exports: $Exports<'eslint/lib/rule-tester/rule-tester'>;
 }
 declare module 'eslint/lib/rules/accessor-pairs.js' {
   declare module.exports: $Exports<'eslint/lib/rules/accessor-pairs'>;
@@ -1645,7 +1803,9 @@ declare module 'eslint/lib/rules/complexity.js' {
   declare module.exports: $Exports<'eslint/lib/rules/complexity'>;
 }
 declare module 'eslint/lib/rules/computed-property-spacing.js' {
-  declare module.exports: $Exports<'eslint/lib/rules/computed-property-spacing'>;
+  declare module.exports: $Exports<
+    'eslint/lib/rules/computed-property-spacing'
+  >;
 }
 declare module 'eslint/lib/rules/consistent-return.js' {
   declare module.exports: $Exports<'eslint/lib/rules/consistent-return'>;
@@ -1661,6 +1821,9 @@ declare module 'eslint/lib/rules/curly.js' {
 }
 declare module 'eslint/lib/rules/default-case.js' {
   declare module.exports: $Exports<'eslint/lib/rules/default-case'>;
+}
+declare module 'eslint/lib/rules/default-param-last.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/default-param-last'>;
 }
 declare module 'eslint/lib/rules/dot-location.js' {
   declare module.exports: $Exports<'eslint/lib/rules/dot-location'>;
@@ -1688,6 +1851,11 @@ declare module 'eslint/lib/rules/func-names.js' {
 }
 declare module 'eslint/lib/rules/func-style.js' {
   declare module.exports: $Exports<'eslint/lib/rules/func-style'>;
+}
+declare module 'eslint/lib/rules/function-call-argument-newline.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/rules/function-call-argument-newline'
+  >;
 }
 declare module 'eslint/lib/rules/function-paren-newline.js' {
   declare module.exports: $Exports<'eslint/lib/rules/function-paren-newline'>;
@@ -1725,6 +1893,12 @@ declare module 'eslint/lib/rules/indent-legacy.js' {
 declare module 'eslint/lib/rules/indent.js' {
   declare module.exports: $Exports<'eslint/lib/rules/indent'>;
 }
+declare module 'eslint/lib/rules/index' {
+  declare module.exports: $Exports<'eslint/lib/rules'>;
+}
+declare module 'eslint/lib/rules/index.js' {
+  declare module.exports: $Exports<'eslint/lib/rules'>;
+}
 declare module 'eslint/lib/rules/init-declarations.js' {
   declare module.exports: $Exports<'eslint/lib/rules/init-declarations'>;
 }
@@ -1750,7 +1924,9 @@ declare module 'eslint/lib/rules/lines-around-directive.js' {
   declare module.exports: $Exports<'eslint/lib/rules/lines-around-directive'>;
 }
 declare module 'eslint/lib/rules/lines-between-class-members.js' {
-  declare module.exports: $Exports<'eslint/lib/rules/lines-between-class-members'>;
+  declare module.exports: $Exports<
+    'eslint/lib/rules/lines-between-class-members'
+  >;
 }
 declare module 'eslint/lib/rules/max-classes-per-file.js' {
   declare module.exports: $Exports<'eslint/lib/rules/max-classes-per-file'>;
@@ -1807,7 +1983,9 @@ declare module 'eslint/lib/rules/no-array-constructor.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-array-constructor'>;
 }
 declare module 'eslint/lib/rules/no-async-promise-executor.js' {
-  declare module.exports: $Exports<'eslint/lib/rules/no-async-promise-executor'>;
+  declare module.exports: $Exports<
+    'eslint/lib/rules/no-async-promise-executor'
+  >;
 }
 declare module 'eslint/lib/rules/no-await-in-loop.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-await-in-loop'>;
@@ -1941,6 +2119,9 @@ declare module 'eslint/lib/rules/no-implicit-globals.js' {
 declare module 'eslint/lib/rules/no-implied-eval.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-implied-eval'>;
 }
+declare module 'eslint/lib/rules/no-import-assign.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/no-import-assign'>;
+}
 declare module 'eslint/lib/rules/no-inline-comments.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-inline-comments'>;
 }
@@ -1978,7 +2159,9 @@ declare module 'eslint/lib/rules/no-magic-numbers.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-magic-numbers'>;
 }
 declare module 'eslint/lib/rules/no-misleading-character-class.js' {
-  declare module.exports: $Exports<'eslint/lib/rules/no-misleading-character-class'>;
+  declare module.exports: $Exports<
+    'eslint/lib/rules/no-misleading-character-class'
+  >;
 }
 declare module 'eslint/lib/rules/no-mixed-operators.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-mixed-operators'>;
@@ -2101,7 +2284,9 @@ declare module 'eslint/lib/rules/no-sequences.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-sequences'>;
 }
 declare module 'eslint/lib/rules/no-shadow-restricted-names.js' {
-  declare module.exports: $Exports<'eslint/lib/rules/no-shadow-restricted-names'>;
+  declare module.exports: $Exports<
+    'eslint/lib/rules/no-shadow-restricted-names'
+  >;
 }
 declare module 'eslint/lib/rules/no-shadow.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-shadow'>;
@@ -2119,7 +2304,9 @@ declare module 'eslint/lib/rules/no-tabs.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-tabs'>;
 }
 declare module 'eslint/lib/rules/no-template-curly-in-string.js' {
-  declare module.exports: $Exports<'eslint/lib/rules/no-template-curly-in-string'>;
+  declare module.exports: $Exports<
+    'eslint/lib/rules/no-template-curly-in-string'
+  >;
 }
 declare module 'eslint/lib/rules/no-ternary.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-ternary'>;
@@ -2149,7 +2336,9 @@ declare module 'eslint/lib/rules/no-unexpected-multiline.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-unexpected-multiline'>;
 }
 declare module 'eslint/lib/rules/no-unmodified-loop-condition.js' {
-  declare module.exports: $Exports<'eslint/lib/rules/no-unmodified-loop-condition'>;
+  declare module.exports: $Exports<
+    'eslint/lib/rules/no-unmodified-loop-condition'
+  >;
 }
 declare module 'eslint/lib/rules/no-unneeded-ternary.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-unneeded-ternary'>;
@@ -2209,13 +2398,17 @@ declare module 'eslint/lib/rules/no-warning-comments.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-warning-comments'>;
 }
 declare module 'eslint/lib/rules/no-whitespace-before-property.js' {
-  declare module.exports: $Exports<'eslint/lib/rules/no-whitespace-before-property'>;
+  declare module.exports: $Exports<
+    'eslint/lib/rules/no-whitespace-before-property'
+  >;
 }
 declare module 'eslint/lib/rules/no-with.js' {
   declare module.exports: $Exports<'eslint/lib/rules/no-with'>;
 }
 declare module 'eslint/lib/rules/nonblock-statement-body-position.js' {
-  declare module.exports: $Exports<'eslint/lib/rules/nonblock-statement-body-position'>;
+  declare module.exports: $Exports<
+    'eslint/lib/rules/nonblock-statement-body-position'
+  >;
 }
 declare module 'eslint/lib/rules/object-curly-newline.js' {
   declare module.exports: $Exports<'eslint/lib/rules/object-curly-newline'>;
@@ -2230,7 +2423,9 @@ declare module 'eslint/lib/rules/object-shorthand.js' {
   declare module.exports: $Exports<'eslint/lib/rules/object-shorthand'>;
 }
 declare module 'eslint/lib/rules/one-var-declaration-per-line.js' {
-  declare module.exports: $Exports<'eslint/lib/rules/one-var-declaration-per-line'>;
+  declare module.exports: $Exports<
+    'eslint/lib/rules/one-var-declaration-per-line'
+  >;
 }
 declare module 'eslint/lib/rules/one-var.js' {
   declare module.exports: $Exports<'eslint/lib/rules/one-var'>;
@@ -2245,7 +2440,9 @@ declare module 'eslint/lib/rules/padded-blocks.js' {
   declare module.exports: $Exports<'eslint/lib/rules/padded-blocks'>;
 }
 declare module 'eslint/lib/rules/padding-line-between-statements.js' {
-  declare module.exports: $Exports<'eslint/lib/rules/padding-line-between-statements'>;
+  declare module.exports: $Exports<
+    'eslint/lib/rules/padding-line-between-statements'
+  >;
 }
 declare module 'eslint/lib/rules/prefer-arrow-callback.js' {
   declare module.exports: $Exports<'eslint/lib/rules/prefer-arrow-callback'>;
@@ -2257,7 +2454,9 @@ declare module 'eslint/lib/rules/prefer-destructuring.js' {
   declare module.exports: $Exports<'eslint/lib/rules/prefer-destructuring'>;
 }
 declare module 'eslint/lib/rules/prefer-named-capture-group.js' {
-  declare module.exports: $Exports<'eslint/lib/rules/prefer-named-capture-group'>;
+  declare module.exports: $Exports<
+    'eslint/lib/rules/prefer-named-capture-group'
+  >;
 }
 declare module 'eslint/lib/rules/prefer-numeric-literals.js' {
   declare module.exports: $Exports<'eslint/lib/rules/prefer-numeric-literals'>;
@@ -2266,10 +2465,15 @@ declare module 'eslint/lib/rules/prefer-object-spread.js' {
   declare module.exports: $Exports<'eslint/lib/rules/prefer-object-spread'>;
 }
 declare module 'eslint/lib/rules/prefer-promise-reject-errors.js' {
-  declare module.exports: $Exports<'eslint/lib/rules/prefer-promise-reject-errors'>;
+  declare module.exports: $Exports<
+    'eslint/lib/rules/prefer-promise-reject-errors'
+  >;
 }
 declare module 'eslint/lib/rules/prefer-reflect.js' {
   declare module.exports: $Exports<'eslint/lib/rules/prefer-reflect'>;
+}
+declare module 'eslint/lib/rules/prefer-regex-literals.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/prefer-regex-literals'>;
 }
 declare module 'eslint/lib/rules/prefer-rest-params.js' {
   declare module.exports: $Exports<'eslint/lib/rules/prefer-rest-params'>;
@@ -2329,7 +2533,9 @@ declare module 'eslint/lib/rules/space-before-blocks.js' {
   declare module.exports: $Exports<'eslint/lib/rules/space-before-blocks'>;
 }
 declare module 'eslint/lib/rules/space-before-function-paren.js' {
-  declare module.exports: $Exports<'eslint/lib/rules/space-before-function-paren'>;
+  declare module.exports: $Exports<
+    'eslint/lib/rules/space-before-function-paren'
+  >;
 }
 declare module 'eslint/lib/rules/space-in-parens.js' {
   declare module.exports: $Exports<'eslint/lib/rules/space-in-parens'>;
@@ -2364,6 +2570,49 @@ declare module 'eslint/lib/rules/unicode-bom.js' {
 declare module 'eslint/lib/rules/use-isnan.js' {
   declare module.exports: $Exports<'eslint/lib/rules/use-isnan'>;
 }
+declare module 'eslint/lib/rules/utils/ast-utils.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/utils/ast-utils'>;
+}
+declare module 'eslint/lib/rules/utils/fix-tracker.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/utils/fix-tracker'>;
+}
+declare module 'eslint/lib/rules/utils/keywords.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/utils/keywords'>;
+}
+declare module 'eslint/lib/rules/utils/lazy-loading-rule-map.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/rules/utils/lazy-loading-rule-map'
+  >;
+}
+declare module 'eslint/lib/rules/utils/patterns/letters.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/utils/patterns/letters'>;
+}
+declare module 'eslint/lib/rules/utils/unicode/index' {
+  declare module.exports: $Exports<'eslint/lib/rules/utils/unicode'>;
+}
+declare module 'eslint/lib/rules/utils/unicode/index.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/utils/unicode'>;
+}
+declare module 'eslint/lib/rules/utils/unicode/is-combining-character.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/rules/utils/unicode/is-combining-character'
+  >;
+}
+declare module 'eslint/lib/rules/utils/unicode/is-emoji-modifier.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/rules/utils/unicode/is-emoji-modifier'
+  >;
+}
+declare module 'eslint/lib/rules/utils/unicode/is-regional-indicator-symbol.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/rules/utils/unicode/is-regional-indicator-symbol'
+  >;
+}
+declare module 'eslint/lib/rules/utils/unicode/is-surrogate-pair.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/rules/utils/unicode/is-surrogate-pair'
+  >;
+}
 declare module 'eslint/lib/rules/valid-jsdoc.js' {
   declare module.exports: $Exports<'eslint/lib/rules/valid-jsdoc'>;
 }
@@ -2385,147 +2634,106 @@ declare module 'eslint/lib/rules/yield-star-spacing.js' {
 declare module 'eslint/lib/rules/yoda.js' {
   declare module.exports: $Exports<'eslint/lib/rules/yoda'>;
 }
-declare module 'eslint/lib/testers/rule-tester.js' {
-  declare module.exports: $Exports<'eslint/lib/testers/rule-tester'>;
+declare module 'eslint/lib/shared/ajv.js' {
+  declare module.exports: $Exports<'eslint/lib/shared/ajv'>;
 }
-declare module 'eslint/lib/token-store/backward-token-comment-cursor.js' {
-  declare module.exports: $Exports<'eslint/lib/token-store/backward-token-comment-cursor'>;
+declare module 'eslint/lib/shared/ast-utils.js' {
+  declare module.exports: $Exports<'eslint/lib/shared/ast-utils'>;
 }
-declare module 'eslint/lib/token-store/backward-token-cursor.js' {
-  declare module.exports: $Exports<'eslint/lib/token-store/backward-token-cursor'>;
+declare module 'eslint/lib/shared/config-ops.js' {
+  declare module.exports: $Exports<'eslint/lib/shared/config-ops'>;
 }
-declare module 'eslint/lib/token-store/cursor.js' {
-  declare module.exports: $Exports<'eslint/lib/token-store/cursor'>;
+declare module 'eslint/lib/shared/config-validator.js' {
+  declare module.exports: $Exports<'eslint/lib/shared/config-validator'>;
 }
-declare module 'eslint/lib/token-store/cursors.js' {
-  declare module.exports: $Exports<'eslint/lib/token-store/cursors'>;
+declare module 'eslint/lib/shared/logging.js' {
+  declare module.exports: $Exports<'eslint/lib/shared/logging'>;
 }
-declare module 'eslint/lib/token-store/decorative-cursor.js' {
-  declare module.exports: $Exports<'eslint/lib/token-store/decorative-cursor'>;
+declare module 'eslint/lib/shared/naming.js' {
+  declare module.exports: $Exports<'eslint/lib/shared/naming'>;
 }
-declare module 'eslint/lib/token-store/filter-cursor.js' {
-  declare module.exports: $Exports<'eslint/lib/token-store/filter-cursor'>;
+declare module 'eslint/lib/shared/relative-module-resolver.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/shared/relative-module-resolver'
+  >;
 }
-declare module 'eslint/lib/token-store/forward-token-comment-cursor.js' {
-  declare module.exports: $Exports<'eslint/lib/token-store/forward-token-comment-cursor'>;
+declare module 'eslint/lib/shared/runtime-info.js' {
+  declare module.exports: $Exports<'eslint/lib/shared/runtime-info'>;
 }
-declare module 'eslint/lib/token-store/forward-token-cursor.js' {
-  declare module.exports: $Exports<'eslint/lib/token-store/forward-token-cursor'>;
+declare module 'eslint/lib/shared/traverser.js' {
+  declare module.exports: $Exports<'eslint/lib/shared/traverser'>;
 }
-declare module 'eslint/lib/token-store/index.js' {
-  declare module.exports: $Exports<'eslint/lib/token-store/index'>;
+declare module 'eslint/lib/shared/types.js' {
+  declare module.exports: $Exports<'eslint/lib/shared/types'>;
 }
-declare module 'eslint/lib/token-store/limit-cursor.js' {
-  declare module.exports: $Exports<'eslint/lib/token-store/limit-cursor'>;
+declare module 'eslint/lib/source-code/index' {
+  declare module.exports: $Exports<'eslint/lib/source-code'>;
 }
-declare module 'eslint/lib/token-store/padded-token-cursor.js' {
-  declare module.exports: $Exports<'eslint/lib/token-store/padded-token-cursor'>;
+declare module 'eslint/lib/source-code/index.js' {
+  declare module.exports: $Exports<'eslint/lib/source-code'>;
 }
-declare module 'eslint/lib/token-store/skip-cursor.js' {
-  declare module.exports: $Exports<'eslint/lib/token-store/skip-cursor'>;
+declare module 'eslint/lib/source-code/source-code.js' {
+  declare module.exports: $Exports<'eslint/lib/source-code/source-code'>;
 }
-declare module 'eslint/lib/token-store/utils.js' {
-  declare module.exports: $Exports<'eslint/lib/token-store/utils'>;
+declare module 'eslint/lib/source-code/token-store/backward-token-comment-cursor.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/source-code/token-store/backward-token-comment-cursor'
+  >;
 }
-declare module 'eslint/lib/util/ajv.js' {
-  declare module.exports: $Exports<'eslint/lib/util/ajv'>;
+declare module 'eslint/lib/source-code/token-store/backward-token-cursor.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/source-code/token-store/backward-token-cursor'
+  >;
 }
-declare module 'eslint/lib/util/apply-disable-directives.js' {
-  declare module.exports: $Exports<'eslint/lib/util/apply-disable-directives'>;
+declare module 'eslint/lib/source-code/token-store/cursor.js' {
+  declare module.exports: $Exports<'eslint/lib/source-code/token-store/cursor'>;
 }
-declare module 'eslint/lib/util/ast-utils.js' {
-  declare module.exports: $Exports<'eslint/lib/util/ast-utils'>;
+declare module 'eslint/lib/source-code/token-store/cursors.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/source-code/token-store/cursors'
+  >;
 }
-declare module 'eslint/lib/util/config-comment-parser.js' {
-  declare module.exports: $Exports<'eslint/lib/util/config-comment-parser'>;
+declare module 'eslint/lib/source-code/token-store/decorative-cursor.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/source-code/token-store/decorative-cursor'
+  >;
 }
-declare module 'eslint/lib/util/file-finder.js' {
-  declare module.exports: $Exports<'eslint/lib/util/file-finder'>;
+declare module 'eslint/lib/source-code/token-store/filter-cursor.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/source-code/token-store/filter-cursor'
+  >;
 }
-declare module 'eslint/lib/util/fix-tracker.js' {
-  declare module.exports: $Exports<'eslint/lib/util/fix-tracker'>;
+declare module 'eslint/lib/source-code/token-store/forward-token-comment-cursor.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/source-code/token-store/forward-token-comment-cursor'
+  >;
 }
-declare module 'eslint/lib/util/glob-utils.js' {
-  declare module.exports: $Exports<'eslint/lib/util/glob-utils'>;
+declare module 'eslint/lib/source-code/token-store/forward-token-cursor.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/source-code/token-store/forward-token-cursor'
+  >;
 }
-declare module 'eslint/lib/util/glob.js' {
-  declare module.exports: $Exports<'eslint/lib/util/glob'>;
+declare module 'eslint/lib/source-code/token-store/index' {
+  declare module.exports: $Exports<'eslint/lib/source-code/token-store'>;
 }
-declare module 'eslint/lib/util/hash.js' {
-  declare module.exports: $Exports<'eslint/lib/util/hash'>;
+declare module 'eslint/lib/source-code/token-store/index.js' {
+  declare module.exports: $Exports<'eslint/lib/source-code/token-store'>;
 }
-declare module 'eslint/lib/util/ignored-paths.js' {
-  declare module.exports: $Exports<'eslint/lib/util/ignored-paths'>;
+declare module 'eslint/lib/source-code/token-store/limit-cursor.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/source-code/token-store/limit-cursor'
+  >;
 }
-declare module 'eslint/lib/util/interpolate.js' {
-  declare module.exports: $Exports<'eslint/lib/util/interpolate'>;
+declare module 'eslint/lib/source-code/token-store/padded-token-cursor.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/source-code/token-store/padded-token-cursor'
+  >;
 }
-declare module 'eslint/lib/util/keywords.js' {
-  declare module.exports: $Exports<'eslint/lib/util/keywords'>;
+declare module 'eslint/lib/source-code/token-store/skip-cursor.js' {
+  declare module.exports: $Exports<
+    'eslint/lib/source-code/token-store/skip-cursor'
+  >;
 }
-declare module 'eslint/lib/util/lint-result-cache.js' {
-  declare module.exports: $Exports<'eslint/lib/util/lint-result-cache'>;
-}
-declare module 'eslint/lib/util/logging.js' {
-  declare module.exports: $Exports<'eslint/lib/util/logging'>;
-}
-declare module 'eslint/lib/util/module-resolver.js' {
-  declare module.exports: $Exports<'eslint/lib/util/module-resolver'>;
-}
-declare module 'eslint/lib/util/naming.js' {
-  declare module.exports: $Exports<'eslint/lib/util/naming'>;
-}
-declare module 'eslint/lib/util/node-event-generator.js' {
-  declare module.exports: $Exports<'eslint/lib/util/node-event-generator'>;
-}
-declare module 'eslint/lib/util/npm-utils.js' {
-  declare module.exports: $Exports<'eslint/lib/util/npm-utils'>;
-}
-declare module 'eslint/lib/util/path-utils.js' {
-  declare module.exports: $Exports<'eslint/lib/util/path-utils'>;
-}
-declare module 'eslint/lib/util/patterns/letters.js' {
-  declare module.exports: $Exports<'eslint/lib/util/patterns/letters'>;
-}
-declare module 'eslint/lib/util/report-translator.js' {
-  declare module.exports: $Exports<'eslint/lib/util/report-translator'>;
-}
-declare module 'eslint/lib/util/rule-fixer.js' {
-  declare module.exports: $Exports<'eslint/lib/util/rule-fixer'>;
-}
-declare module 'eslint/lib/util/safe-emitter.js' {
-  declare module.exports: $Exports<'eslint/lib/util/safe-emitter'>;
-}
-declare module 'eslint/lib/util/source-code-fixer.js' {
-  declare module.exports: $Exports<'eslint/lib/util/source-code-fixer'>;
-}
-declare module 'eslint/lib/util/source-code-utils.js' {
-  declare module.exports: $Exports<'eslint/lib/util/source-code-utils'>;
-}
-declare module 'eslint/lib/util/source-code.js' {
-  declare module.exports: $Exports<'eslint/lib/util/source-code'>;
-}
-declare module 'eslint/lib/util/timing.js' {
-  declare module.exports: $Exports<'eslint/lib/util/timing'>;
-}
-declare module 'eslint/lib/util/traverser.js' {
-  declare module.exports: $Exports<'eslint/lib/util/traverser'>;
-}
-declare module 'eslint/lib/util/unicode/index.js' {
-  declare module.exports: $Exports<'eslint/lib/util/unicode/index'>;
-}
-declare module 'eslint/lib/util/unicode/is-combining-character.js' {
-  declare module.exports: $Exports<'eslint/lib/util/unicode/is-combining-character'>;
-}
-declare module 'eslint/lib/util/unicode/is-emoji-modifier.js' {
-  declare module.exports: $Exports<'eslint/lib/util/unicode/is-emoji-modifier'>;
-}
-declare module 'eslint/lib/util/unicode/is-regional-indicator-symbol.js' {
-  declare module.exports: $Exports<'eslint/lib/util/unicode/is-regional-indicator-symbol'>;
-}
-declare module 'eslint/lib/util/unicode/is-surrogate-pair.js' {
-  declare module.exports: $Exports<'eslint/lib/util/unicode/is-surrogate-pair'>;
-}
-declare module 'eslint/lib/util/xml-escape.js' {
-  declare module.exports: $Exports<'eslint/lib/util/xml-escape'>;
+declare module 'eslint/lib/source-code/token-store/utils.js' {
+  declare module.exports: $Exports<'eslint/lib/source-code/token-store/utils'>;
 }
