@@ -1,5 +1,5 @@
-// flow-typed signature: 092ddb0530b7210a4dd7c20a438c80d3
-// flow-typed version: <<STUB>>/flow-typed_v2.5.2/flow_v0.99.1
+// flow-typed signature: 8cdc50e62d557e2f5a00fc7bc55176ea
+// flow-typed version: <<STUB>>/flow-typed_v2.6.2/flow_v0.112.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -51,10 +51,6 @@ declare module 'flow-typed/dist/commands/update' {
 }
 
 declare module 'flow-typed/dist/commands/validateDefs' {
-  declare module.exports: any;
-}
-
-declare module 'flow-typed/dist/commands/version' {
   declare module.exports: any;
 }
 
@@ -142,9 +138,6 @@ declare module 'flow-typed/dist/commands/update.js' {
 }
 declare module 'flow-typed/dist/commands/validateDefs.js' {
   declare module.exports: $Exports<'flow-typed/dist/commands/validateDefs'>;
-}
-declare module 'flow-typed/dist/commands/version.js' {
-  declare module.exports: $Exports<'flow-typed/dist/commands/version'>;
 }
 declare module 'flow-typed/dist/lib/cacheRepoUtils.js' {
   declare module.exports: $Exports<'flow-typed/dist/lib/cacheRepoUtils'>;
