@@ -1,5 +1,5 @@
-// flow-typed signature: 16c393ad8a3fe2878254b66f4d8b14f8
-// flow-typed version: <<STUB>>/next_v8.1.0/flow_v0.99.1
+// flow-typed signature: 084221ab01c32e988ced1f2e588a2da1
+// flow-typed version: <<STUB>>/next_v9.1.3/flow_v0.112.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -58,7 +58,19 @@ declare module 'next/dist/build/babel/plugins/commonjs' {
   declare module.exports: any;
 }
 
+declare module 'next/dist/build/babel/plugins/jsx-pragma' {
+  declare module.exports: any;
+}
+
 declare module 'next/dist/build/babel/plugins/next-data' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/build/babel/plugins/next-page-config' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/build/babel/plugins/optimize-hook-destructuring' {
   declare module.exports: any;
 }
 
@@ -78,15 +90,11 @@ declare module 'next/dist/build/entries' {
   declare module.exports: any;
 }
 
-declare module 'next/dist/build/flying-shuttle' {
-  declare module.exports: any;
-}
-
 declare module 'next/dist/build/generate-build-id' {
   declare module.exports: any;
 }
 
-declare module 'next/dist/build/index' {
+declare module 'next/dist/build' {
   declare module.exports: any;
 }
 
@@ -94,7 +102,7 @@ declare module 'next/dist/build/is-writeable' {
   declare module.exports: any;
 }
 
-declare module 'next/dist/build/output/index' {
+declare module 'next/dist/build/output' {
   declare module.exports: any;
 }
 
@@ -103,6 +111,22 @@ declare module 'next/dist/build/output/log' {
 }
 
 declare module 'next/dist/build/output/store' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/build/plugins/collect-plugins' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/build/polyfills/fetch' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/build/profiler/profiler' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/build/spinner' {
   declare module.exports: any;
 }
 
@@ -118,6 +142,10 @@ declare module 'next/dist/build/webpack/loaders/emit-file-loader' {
   declare module.exports: any;
 }
 
+declare module 'next/dist/build/webpack/loaders/error-loader' {
+  declare module.exports: any;
+}
+
 declare module 'next/dist/build/webpack/loaders/next-babel-loader' {
   declare module.exports: any;
 }
@@ -130,7 +158,7 @@ declare module 'next/dist/build/webpack/loaders/next-data-loader' {
   declare module.exports: any;
 }
 
-declare module 'next/dist/build/webpack/loaders/next-minify-loader' {
+declare module 'next/dist/build/webpack/loaders/next-plugin-loader' {
   declare module.exports: any;
 }
 
@@ -142,15 +170,7 @@ declare module 'next/dist/build/webpack/loaders/noop-loader' {
   declare module.exports: any;
 }
 
-declare module 'next/dist/build/webpack/plugins/all-modules-identified-plugin' {
-  declare module.exports: any;
-}
-
 declare module 'next/dist/build/webpack/plugins/build-manifest-plugin' {
-  declare module.exports: any;
-}
-
-declare module 'next/dist/build/webpack/plugins/chunk-graph-plugin' {
   declare module.exports: any;
 }
 
@@ -158,7 +178,19 @@ declare module 'next/dist/build/webpack/plugins/chunk-names-plugin' {
   declare module.exports: any;
 }
 
-declare module 'next/dist/build/webpack/plugins/hashed-chunk-ids-plugin' {
+declare module 'next/dist/build/webpack/plugins/css-minimizer-plugin' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/build/webpack/plugins/dll-import' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/build/webpack/plugins/next-drop-client-page-plugin' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/build/webpack/plugins/next-esm-plugin' {
   declare module.exports: any;
 }
 
@@ -178,6 +210,10 @@ declare module 'next/dist/build/webpack/plugins/pages-manifest-plugin' {
   declare module.exports: any;
 }
 
+declare module 'next/dist/build/webpack/plugins/profiling-plugin' {
+  declare module.exports: any;
+}
+
 declare module 'next/dist/build/webpack/plugins/react-loadable-plugin' {
   declare module.exports: any;
 }
@@ -186,11 +222,7 @@ declare module 'next/dist/build/webpack/plugins/serverless-plugin' {
   declare module.exports: any;
 }
 
-declare module 'next/dist/build/webpack/plugins/shared-runtime-plugin' {
-  declare module.exports: any;
-}
-
-declare module 'next/dist/build/webpack/plugins/terser-webpack-plugin/src/index' {
+declare module 'next/dist/build/webpack/plugins/terser-webpack-plugin/src' {
   declare module.exports: any;
 }
 
@@ -199,10 +231,6 @@ declare module 'next/dist/build/webpack/plugins/terser-webpack-plugin/src/minify
 }
 
 declare module 'next/dist/build/webpack/plugins/terser-webpack-plugin/src/TaskRunner' {
-  declare module.exports: any;
-}
-
-declare module 'next/dist/build/webpack/plugins/terser-webpack-plugin/src/worker' {
   declare module.exports: any;
 }
 
@@ -230,27 +258,55 @@ declare module 'next/dist/cli/next-start' {
   declare module.exports: any;
 }
 
-declare module 'next/dist/client/amp-dev' {
+declare module 'next/dist/cli/next-telemetry' {
   declare module.exports: any;
 }
 
-declare module 'next/dist/client/dev-error-overlay/eventsource' {
+declare module 'next/dist/client/dev/amp-dev' {
   declare module.exports: any;
 }
 
-declare module 'next/dist/client/dev-error-overlay/format-webpack-messages' {
+declare module 'next/dist/client/dev/dev-build-watcher' {
   declare module.exports: any;
 }
 
-declare module 'next/dist/client/dev-error-overlay/hot-dev-client' {
+declare module 'next/dist/client/dev/error-overlay/eventsource' {
   declare module.exports: any;
 }
 
-declare module 'next/dist/client/error-boundary' {
+declare module 'next/dist/client/dev/error-overlay/format-webpack-messages' {
   declare module.exports: any;
 }
 
-declare module 'next/dist/client/event-source-polyfill' {
+declare module 'next/dist/client/dev/error-overlay/hot-dev-client' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/client/dev/error-overlay/source-map-support' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/client/dev/event-source-polyfill' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/client/dev/noop' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/client/dev/on-demand-entries-client' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/client/dev/on-demand-entries-utils' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/client/dev/prerender-indicator' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/client/dev/webpack-hot-middleware-client' {
   declare module.exports: any;
 }
 
@@ -258,7 +314,7 @@ declare module 'next/dist/client/head-manager' {
   declare module.exports: any;
 }
 
-declare module 'next/dist/client/index' {
+declare module 'next/dist/client' {
   declare module.exports: any;
 }
 
@@ -274,19 +330,11 @@ declare module 'next/dist/client/next' {
   declare module.exports: any;
 }
 
-declare module 'next/dist/client/noop' {
-  declare module.exports: any;
-}
-
-declare module 'next/dist/client/on-demand-entries-client' {
-  declare module.exports: any;
-}
-
-declare module 'next/dist/client/on-demand-entries-utils' {
-  declare module.exports: any;
-}
-
 declare module 'next/dist/client/page-loader' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/client/polyfills' {
   declare module.exports: any;
 }
 
@@ -294,31 +342,23 @@ declare module 'next/dist/client/router' {
   declare module.exports: any;
 }
 
-declare module 'next/dist/client/source-map-support' {
-  declare module.exports: any;
-}
-
-declare module 'next/dist/client/webpack-hot-middleware-client' {
-  declare module.exports: any;
-}
-
 declare module 'next/dist/client/with-router' {
   declare module.exports: any;
 }
 
-declare module 'next/dist/compiled/arg/index' {
+declare module 'next/dist/compiled/arg' {
   declare module.exports: any;
 }
 
-declare module 'next/dist/compiled/nanoid/index' {
+declare module 'next/dist/compiled/nanoid' {
   declare module.exports: any;
 }
 
-declare module 'next/dist/compiled/resolve/index' {
+declare module 'next/dist/compiled/resolve' {
   declare module.exports: any;
 }
 
-declare module 'next/dist/compiled/text-table/index' {
+declare module 'next/dist/compiled/text-table' {
   declare module.exports: any;
 }
 
@@ -326,11 +366,15 @@ declare module 'next/dist/compiled/unistore/unistore' {
   declare module.exports: any;
 }
 
-declare module 'next/dist/export/index' {
+declare module 'next/dist/export' {
   declare module.exports: any;
 }
 
 declare module 'next/dist/export/worker' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/lib/coalesced-function' {
   declare module.exports: any;
 }
 
@@ -342,11 +386,195 @@ declare module 'next/dist/lib/data' {
   declare module.exports: any;
 }
 
+declare module 'next/dist/lib/file-exists' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/lib/find-pages-dir' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/lib/pretty-bytes' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/lib/recursive-copy' {
+  declare module.exports: any;
+}
+
 declare module 'next/dist/lib/recursive-delete' {
   declare module.exports: any;
 }
 
 declare module 'next/dist/lib/recursive-readdir' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/lib/resolve-request' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/lib/verifyTypeScriptSetup' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/lib/amp-context' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/lib/amp' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/lib/constants' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/lib/data-manager-context' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/lib/data-manager' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/lib/document-context' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/lib/dynamic' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/lib/head-manager-context' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/lib/head' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/lib/loadable-context' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/lib/loadable' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/lib/mitt' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/lib/request-context' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/lib/router-context' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/lib/router/rewrite-url-for-export' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/lib/router/router' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/lib/router/utils' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/lib/router/utils/is-dynamic' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/lib/router/utils/route-matcher' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/lib/router/utils/route-regex' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/lib/router/utils/sorted-routes' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/lib/runtime-config' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/lib/side-effect' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/lib/utils' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/server/api-utils' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/server/config' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/server/get-page-files' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/server/lib/path-match' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/server/lib/recursive-readdir-sync' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/server/load-components' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/server/next-server' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/server/normalize-page-path' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/server/optimize-amp' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/server/render' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/server/require' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/server/router' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/server/send-html' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/server/serve-static' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/server/spr-cache' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/next-server/server/utils' {
   declare module.exports: any;
 }
 
@@ -402,6 +630,42 @@ declare module 'next/dist/server/on-demand-entry-handler' {
   declare module.exports: any;
 }
 
+declare module 'next/dist/telemetry/anonymous-meta' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/telemetry/ci-info' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/telemetry/events/build' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/telemetry/events' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/telemetry/events/plugins' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/telemetry/events/version' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/telemetry/post-payload' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/telemetry/project-id' {
+  declare module.exports: any;
+}
+
+declare module 'next/dist/telemetry/storage' {
+  declare module.exports: any;
+}
+
 declare module 'next/document' {
   declare module.exports: any;
 }
@@ -449,16 +713,33 @@ declare module 'next/data.js' {
   declare module.exports: $Exports<'next/data'>;
 }
 declare module 'next/dist/build/babel/plugins/amp-attributes.js' {
-  declare module.exports: $Exports<'next/dist/build/babel/plugins/amp-attributes'>;
+  declare module.exports: $Exports<
+    'next/dist/build/babel/plugins/amp-attributes'
+  >;
 }
 declare module 'next/dist/build/babel/plugins/commonjs.js' {
   declare module.exports: $Exports<'next/dist/build/babel/plugins/commonjs'>;
 }
+declare module 'next/dist/build/babel/plugins/jsx-pragma.js' {
+  declare module.exports: $Exports<'next/dist/build/babel/plugins/jsx-pragma'>;
+}
 declare module 'next/dist/build/babel/plugins/next-data.js' {
   declare module.exports: $Exports<'next/dist/build/babel/plugins/next-data'>;
 }
+declare module 'next/dist/build/babel/plugins/next-page-config.js' {
+  declare module.exports: $Exports<
+    'next/dist/build/babel/plugins/next-page-config'
+  >;
+}
+declare module 'next/dist/build/babel/plugins/optimize-hook-destructuring.js' {
+  declare module.exports: $Exports<
+    'next/dist/build/babel/plugins/optimize-hook-destructuring'
+  >;
+}
 declare module 'next/dist/build/babel/plugins/react-loadable-plugin.js' {
-  declare module.exports: $Exports<'next/dist/build/babel/plugins/react-loadable-plugin'>;
+  declare module.exports: $Exports<
+    'next/dist/build/babel/plugins/react-loadable-plugin'
+  >;
 }
 declare module 'next/dist/build/babel/preset.js' {
   declare module.exports: $Exports<'next/dist/build/babel/preset'>;
@@ -469,26 +750,41 @@ declare module 'next/dist/build/compiler.js' {
 declare module 'next/dist/build/entries.js' {
   declare module.exports: $Exports<'next/dist/build/entries'>;
 }
-declare module 'next/dist/build/flying-shuttle.js' {
-  declare module.exports: $Exports<'next/dist/build/flying-shuttle'>;
-}
 declare module 'next/dist/build/generate-build-id.js' {
   declare module.exports: $Exports<'next/dist/build/generate-build-id'>;
 }
+declare module 'next/dist/build/index' {
+  declare module.exports: $Exports<'next/dist/build'>;
+}
 declare module 'next/dist/build/index.js' {
-  declare module.exports: $Exports<'next/dist/build/index'>;
+  declare module.exports: $Exports<'next/dist/build'>;
 }
 declare module 'next/dist/build/is-writeable.js' {
   declare module.exports: $Exports<'next/dist/build/is-writeable'>;
 }
+declare module 'next/dist/build/output/index' {
+  declare module.exports: $Exports<'next/dist/build/output'>;
+}
 declare module 'next/dist/build/output/index.js' {
-  declare module.exports: $Exports<'next/dist/build/output/index'>;
+  declare module.exports: $Exports<'next/dist/build/output'>;
 }
 declare module 'next/dist/build/output/log.js' {
   declare module.exports: $Exports<'next/dist/build/output/log'>;
 }
 declare module 'next/dist/build/output/store.js' {
   declare module.exports: $Exports<'next/dist/build/output/store'>;
+}
+declare module 'next/dist/build/plugins/collect-plugins.js' {
+  declare module.exports: $Exports<'next/dist/build/plugins/collect-plugins'>;
+}
+declare module 'next/dist/build/polyfills/fetch.js' {
+  declare module.exports: $Exports<'next/dist/build/polyfills/fetch'>;
+}
+declare module 'next/dist/build/profiler/profiler.js' {
+  declare module.exports: $Exports<'next/dist/build/profiler/profiler'>;
+}
+declare module 'next/dist/build/spinner.js' {
+  declare module.exports: $Exports<'next/dist/build/spinner'>;
 }
 declare module 'next/dist/build/utils.js' {
   declare module.exports: $Exports<'next/dist/build/utils'>;
@@ -497,76 +793,134 @@ declare module 'next/dist/build/webpack-config.js' {
   declare module.exports: $Exports<'next/dist/build/webpack-config'>;
 }
 declare module 'next/dist/build/webpack/loaders/emit-file-loader.js' {
-  declare module.exports: $Exports<'next/dist/build/webpack/loaders/emit-file-loader'>;
+  declare module.exports: $Exports<
+    'next/dist/build/webpack/loaders/emit-file-loader'
+  >;
+}
+declare module 'next/dist/build/webpack/loaders/error-loader.js' {
+  declare module.exports: $Exports<
+    'next/dist/build/webpack/loaders/error-loader'
+  >;
 }
 declare module 'next/dist/build/webpack/loaders/next-babel-loader.js' {
-  declare module.exports: $Exports<'next/dist/build/webpack/loaders/next-babel-loader'>;
+  declare module.exports: $Exports<
+    'next/dist/build/webpack/loaders/next-babel-loader'
+  >;
 }
 declare module 'next/dist/build/webpack/loaders/next-client-pages-loader.js' {
-  declare module.exports: $Exports<'next/dist/build/webpack/loaders/next-client-pages-loader'>;
+  declare module.exports: $Exports<
+    'next/dist/build/webpack/loaders/next-client-pages-loader'
+  >;
 }
 declare module 'next/dist/build/webpack/loaders/next-data-loader.js' {
-  declare module.exports: $Exports<'next/dist/build/webpack/loaders/next-data-loader'>;
+  declare module.exports: $Exports<
+    'next/dist/build/webpack/loaders/next-data-loader'
+  >;
 }
-declare module 'next/dist/build/webpack/loaders/next-minify-loader.js' {
-  declare module.exports: $Exports<'next/dist/build/webpack/loaders/next-minify-loader'>;
+declare module 'next/dist/build/webpack/loaders/next-plugin-loader.js' {
+  declare module.exports: $Exports<
+    'next/dist/build/webpack/loaders/next-plugin-loader'
+  >;
 }
 declare module 'next/dist/build/webpack/loaders/next-serverless-loader.js' {
-  declare module.exports: $Exports<'next/dist/build/webpack/loaders/next-serverless-loader'>;
+  declare module.exports: $Exports<
+    'next/dist/build/webpack/loaders/next-serverless-loader'
+  >;
 }
 declare module 'next/dist/build/webpack/loaders/noop-loader.js' {
-  declare module.exports: $Exports<'next/dist/build/webpack/loaders/noop-loader'>;
-}
-declare module 'next/dist/build/webpack/plugins/all-modules-identified-plugin.js' {
-  declare module.exports: $Exports<'next/dist/build/webpack/plugins/all-modules-identified-plugin'>;
+  declare module.exports: $Exports<
+    'next/dist/build/webpack/loaders/noop-loader'
+  >;
 }
 declare module 'next/dist/build/webpack/plugins/build-manifest-plugin.js' {
-  declare module.exports: $Exports<'next/dist/build/webpack/plugins/build-manifest-plugin'>;
-}
-declare module 'next/dist/build/webpack/plugins/chunk-graph-plugin.js' {
-  declare module.exports: $Exports<'next/dist/build/webpack/plugins/chunk-graph-plugin'>;
+  declare module.exports: $Exports<
+    'next/dist/build/webpack/plugins/build-manifest-plugin'
+  >;
 }
 declare module 'next/dist/build/webpack/plugins/chunk-names-plugin.js' {
-  declare module.exports: $Exports<'next/dist/build/webpack/plugins/chunk-names-plugin'>;
+  declare module.exports: $Exports<
+    'next/dist/build/webpack/plugins/chunk-names-plugin'
+  >;
 }
-declare module 'next/dist/build/webpack/plugins/hashed-chunk-ids-plugin.js' {
-  declare module.exports: $Exports<'next/dist/build/webpack/plugins/hashed-chunk-ids-plugin'>;
+declare module 'next/dist/build/webpack/plugins/css-minimizer-plugin.js' {
+  declare module.exports: $Exports<
+    'next/dist/build/webpack/plugins/css-minimizer-plugin'
+  >;
+}
+declare module 'next/dist/build/webpack/plugins/dll-import.js' {
+  declare module.exports: $Exports<
+    'next/dist/build/webpack/plugins/dll-import'
+  >;
+}
+declare module 'next/dist/build/webpack/plugins/next-drop-client-page-plugin.js' {
+  declare module.exports: $Exports<
+    'next/dist/build/webpack/plugins/next-drop-client-page-plugin'
+  >;
+}
+declare module 'next/dist/build/webpack/plugins/next-esm-plugin.js' {
+  declare module.exports: $Exports<
+    'next/dist/build/webpack/plugins/next-esm-plugin'
+  >;
 }
 declare module 'next/dist/build/webpack/plugins/nextjs-require-cache-hot-reloader.js' {
-  declare module.exports: $Exports<'next/dist/build/webpack/plugins/nextjs-require-cache-hot-reloader'>;
+  declare module.exports: $Exports<
+    'next/dist/build/webpack/plugins/nextjs-require-cache-hot-reloader'
+  >;
 }
 declare module 'next/dist/build/webpack/plugins/nextjs-ssr-import.js' {
-  declare module.exports: $Exports<'next/dist/build/webpack/plugins/nextjs-ssr-import'>;
+  declare module.exports: $Exports<
+    'next/dist/build/webpack/plugins/nextjs-ssr-import'
+  >;
 }
 declare module 'next/dist/build/webpack/plugins/nextjs-ssr-module-cache.js' {
-  declare module.exports: $Exports<'next/dist/build/webpack/plugins/nextjs-ssr-module-cache'>;
+  declare module.exports: $Exports<
+    'next/dist/build/webpack/plugins/nextjs-ssr-module-cache'
+  >;
 }
 declare module 'next/dist/build/webpack/plugins/pages-manifest-plugin.js' {
-  declare module.exports: $Exports<'next/dist/build/webpack/plugins/pages-manifest-plugin'>;
+  declare module.exports: $Exports<
+    'next/dist/build/webpack/plugins/pages-manifest-plugin'
+  >;
+}
+declare module 'next/dist/build/webpack/plugins/profiling-plugin.js' {
+  declare module.exports: $Exports<
+    'next/dist/build/webpack/plugins/profiling-plugin'
+  >;
 }
 declare module 'next/dist/build/webpack/plugins/react-loadable-plugin.js' {
-  declare module.exports: $Exports<'next/dist/build/webpack/plugins/react-loadable-plugin'>;
+  declare module.exports: $Exports<
+    'next/dist/build/webpack/plugins/react-loadable-plugin'
+  >;
 }
 declare module 'next/dist/build/webpack/plugins/serverless-plugin.js' {
-  declare module.exports: $Exports<'next/dist/build/webpack/plugins/serverless-plugin'>;
+  declare module.exports: $Exports<
+    'next/dist/build/webpack/plugins/serverless-plugin'
+  >;
 }
-declare module 'next/dist/build/webpack/plugins/shared-runtime-plugin.js' {
-  declare module.exports: $Exports<'next/dist/build/webpack/plugins/shared-runtime-plugin'>;
+declare module 'next/dist/build/webpack/plugins/terser-webpack-plugin/src/index' {
+  declare module.exports: $Exports<
+    'next/dist/build/webpack/plugins/terser-webpack-plugin/src'
+  >;
 }
 declare module 'next/dist/build/webpack/plugins/terser-webpack-plugin/src/index.js' {
-  declare module.exports: $Exports<'next/dist/build/webpack/plugins/terser-webpack-plugin/src/index'>;
+  declare module.exports: $Exports<
+    'next/dist/build/webpack/plugins/terser-webpack-plugin/src'
+  >;
 }
 declare module 'next/dist/build/webpack/plugins/terser-webpack-plugin/src/minify.js' {
-  declare module.exports: $Exports<'next/dist/build/webpack/plugins/terser-webpack-plugin/src/minify'>;
+  declare module.exports: $Exports<
+    'next/dist/build/webpack/plugins/terser-webpack-plugin/src/minify'
+  >;
 }
 declare module 'next/dist/build/webpack/plugins/terser-webpack-plugin/src/TaskRunner.js' {
-  declare module.exports: $Exports<'next/dist/build/webpack/plugins/terser-webpack-plugin/src/TaskRunner'>;
-}
-declare module 'next/dist/build/webpack/plugins/terser-webpack-plugin/src/worker.js' {
-  declare module.exports: $Exports<'next/dist/build/webpack/plugins/terser-webpack-plugin/src/worker'>;
+  declare module.exports: $Exports<
+    'next/dist/build/webpack/plugins/terser-webpack-plugin/src/TaskRunner'
+  >;
 }
 declare module 'next/dist/build/webpack/plugins/unlink-removed-pages-plugin.js' {
-  declare module.exports: $Exports<'next/dist/build/webpack/plugins/unlink-removed-pages-plugin'>;
+  declare module.exports: $Exports<
+    'next/dist/build/webpack/plugins/unlink-removed-pages-plugin'
+  >;
 }
 declare module 'next/dist/build/write-build-id.js' {
   declare module.exports: $Exports<'next/dist/build/write-build-id'>;
@@ -583,29 +937,69 @@ declare module 'next/dist/cli/next-export.js' {
 declare module 'next/dist/cli/next-start.js' {
   declare module.exports: $Exports<'next/dist/cli/next-start'>;
 }
-declare module 'next/dist/client/amp-dev.js' {
-  declare module.exports: $Exports<'next/dist/client/amp-dev'>;
+declare module 'next/dist/cli/next-telemetry.js' {
+  declare module.exports: $Exports<'next/dist/cli/next-telemetry'>;
 }
-declare module 'next/dist/client/dev-error-overlay/eventsource.js' {
-  declare module.exports: $Exports<'next/dist/client/dev-error-overlay/eventsource'>;
+declare module 'next/dist/client/dev/amp-dev.js' {
+  declare module.exports: $Exports<'next/dist/client/dev/amp-dev'>;
 }
-declare module 'next/dist/client/dev-error-overlay/format-webpack-messages.js' {
-  declare module.exports: $Exports<'next/dist/client/dev-error-overlay/format-webpack-messages'>;
+declare module 'next/dist/client/dev/dev-build-watcher.js' {
+  declare module.exports: $Exports<'next/dist/client/dev/dev-build-watcher'>;
 }
-declare module 'next/dist/client/dev-error-overlay/hot-dev-client.js' {
-  declare module.exports: $Exports<'next/dist/client/dev-error-overlay/hot-dev-client'>;
+declare module 'next/dist/client/dev/error-overlay/eventsource.js' {
+  declare module.exports: $Exports<
+    'next/dist/client/dev/error-overlay/eventsource'
+  >;
 }
-declare module 'next/dist/client/error-boundary.js' {
-  declare module.exports: $Exports<'next/dist/client/error-boundary'>;
+declare module 'next/dist/client/dev/error-overlay/format-webpack-messages.js' {
+  declare module.exports: $Exports<
+    'next/dist/client/dev/error-overlay/format-webpack-messages'
+  >;
 }
-declare module 'next/dist/client/event-source-polyfill.js' {
-  declare module.exports: $Exports<'next/dist/client/event-source-polyfill'>;
+declare module 'next/dist/client/dev/error-overlay/hot-dev-client.js' {
+  declare module.exports: $Exports<
+    'next/dist/client/dev/error-overlay/hot-dev-client'
+  >;
+}
+declare module 'next/dist/client/dev/error-overlay/source-map-support.js' {
+  declare module.exports: $Exports<
+    'next/dist/client/dev/error-overlay/source-map-support'
+  >;
+}
+declare module 'next/dist/client/dev/event-source-polyfill.js' {
+  declare module.exports: $Exports<
+    'next/dist/client/dev/event-source-polyfill'
+  >;
+}
+declare module 'next/dist/client/dev/noop.js' {
+  declare module.exports: $Exports<'next/dist/client/dev/noop'>;
+}
+declare module 'next/dist/client/dev/on-demand-entries-client.js' {
+  declare module.exports: $Exports<
+    'next/dist/client/dev/on-demand-entries-client'
+  >;
+}
+declare module 'next/dist/client/dev/on-demand-entries-utils.js' {
+  declare module.exports: $Exports<
+    'next/dist/client/dev/on-demand-entries-utils'
+  >;
+}
+declare module 'next/dist/client/dev/prerender-indicator.js' {
+  declare module.exports: $Exports<'next/dist/client/dev/prerender-indicator'>;
+}
+declare module 'next/dist/client/dev/webpack-hot-middleware-client.js' {
+  declare module.exports: $Exports<
+    'next/dist/client/dev/webpack-hot-middleware-client'
+  >;
 }
 declare module 'next/dist/client/head-manager.js' {
   declare module.exports: $Exports<'next/dist/client/head-manager'>;
 }
+declare module 'next/dist/client/index' {
+  declare module.exports: $Exports<'next/dist/client'>;
+}
 declare module 'next/dist/client/index.js' {
-  declare module.exports: $Exports<'next/dist/client/index'>;
+  declare module.exports: $Exports<'next/dist/client'>;
 }
 declare module 'next/dist/client/link.js' {
   declare module.exports: $Exports<'next/dist/client/link'>;
@@ -616,50 +1010,56 @@ declare module 'next/dist/client/next-dev.js' {
 declare module 'next/dist/client/next.js' {
   declare module.exports: $Exports<'next/dist/client/next'>;
 }
-declare module 'next/dist/client/noop.js' {
-  declare module.exports: $Exports<'next/dist/client/noop'>;
-}
-declare module 'next/dist/client/on-demand-entries-client.js' {
-  declare module.exports: $Exports<'next/dist/client/on-demand-entries-client'>;
-}
-declare module 'next/dist/client/on-demand-entries-utils.js' {
-  declare module.exports: $Exports<'next/dist/client/on-demand-entries-utils'>;
-}
 declare module 'next/dist/client/page-loader.js' {
   declare module.exports: $Exports<'next/dist/client/page-loader'>;
+}
+declare module 'next/dist/client/polyfills.js' {
+  declare module.exports: $Exports<'next/dist/client/polyfills'>;
 }
 declare module 'next/dist/client/router.js' {
   declare module.exports: $Exports<'next/dist/client/router'>;
 }
-declare module 'next/dist/client/source-map-support.js' {
-  declare module.exports: $Exports<'next/dist/client/source-map-support'>;
-}
-declare module 'next/dist/client/webpack-hot-middleware-client.js' {
-  declare module.exports: $Exports<'next/dist/client/webpack-hot-middleware-client'>;
-}
 declare module 'next/dist/client/with-router.js' {
   declare module.exports: $Exports<'next/dist/client/with-router'>;
 }
+declare module 'next/dist/compiled/arg/index' {
+  declare module.exports: $Exports<'next/dist/compiled/arg'>;
+}
 declare module 'next/dist/compiled/arg/index.js' {
-  declare module.exports: $Exports<'next/dist/compiled/arg/index'>;
+  declare module.exports: $Exports<'next/dist/compiled/arg'>;
+}
+declare module 'next/dist/compiled/nanoid/index' {
+  declare module.exports: $Exports<'next/dist/compiled/nanoid'>;
 }
 declare module 'next/dist/compiled/nanoid/index.js' {
-  declare module.exports: $Exports<'next/dist/compiled/nanoid/index'>;
+  declare module.exports: $Exports<'next/dist/compiled/nanoid'>;
+}
+declare module 'next/dist/compiled/resolve/index' {
+  declare module.exports: $Exports<'next/dist/compiled/resolve'>;
 }
 declare module 'next/dist/compiled/resolve/index.js' {
-  declare module.exports: $Exports<'next/dist/compiled/resolve/index'>;
+  declare module.exports: $Exports<'next/dist/compiled/resolve'>;
+}
+declare module 'next/dist/compiled/text-table/index' {
+  declare module.exports: $Exports<'next/dist/compiled/text-table'>;
 }
 declare module 'next/dist/compiled/text-table/index.js' {
-  declare module.exports: $Exports<'next/dist/compiled/text-table/index'>;
+  declare module.exports: $Exports<'next/dist/compiled/text-table'>;
 }
 declare module 'next/dist/compiled/unistore/unistore.js' {
   declare module.exports: $Exports<'next/dist/compiled/unistore/unistore'>;
 }
+declare module 'next/dist/export/index' {
+  declare module.exports: $Exports<'next/dist/export'>;
+}
 declare module 'next/dist/export/index.js' {
-  declare module.exports: $Exports<'next/dist/export/index'>;
+  declare module.exports: $Exports<'next/dist/export'>;
 }
 declare module 'next/dist/export/worker.js' {
   declare module.exports: $Exports<'next/dist/export/worker'>;
+}
+declare module 'next/dist/lib/coalesced-function.js' {
+  declare module.exports: $Exports<'next/dist/lib/coalesced-function'>;
 }
 declare module 'next/dist/lib/constants.js' {
   declare module.exports: $Exports<'next/dist/lib/constants'>;
@@ -667,11 +1067,180 @@ declare module 'next/dist/lib/constants.js' {
 declare module 'next/dist/lib/data.js' {
   declare module.exports: $Exports<'next/dist/lib/data'>;
 }
+declare module 'next/dist/lib/file-exists.js' {
+  declare module.exports: $Exports<'next/dist/lib/file-exists'>;
+}
+declare module 'next/dist/lib/find-pages-dir.js' {
+  declare module.exports: $Exports<'next/dist/lib/find-pages-dir'>;
+}
+declare module 'next/dist/lib/pretty-bytes.js' {
+  declare module.exports: $Exports<'next/dist/lib/pretty-bytes'>;
+}
+declare module 'next/dist/lib/recursive-copy.js' {
+  declare module.exports: $Exports<'next/dist/lib/recursive-copy'>;
+}
 declare module 'next/dist/lib/recursive-delete.js' {
   declare module.exports: $Exports<'next/dist/lib/recursive-delete'>;
 }
 declare module 'next/dist/lib/recursive-readdir.js' {
   declare module.exports: $Exports<'next/dist/lib/recursive-readdir'>;
+}
+declare module 'next/dist/lib/resolve-request.js' {
+  declare module.exports: $Exports<'next/dist/lib/resolve-request'>;
+}
+declare module 'next/dist/lib/verifyTypeScriptSetup.js' {
+  declare module.exports: $Exports<'next/dist/lib/verifyTypeScriptSetup'>;
+}
+declare module 'next/dist/next-server/lib/amp-context.js' {
+  declare module.exports: $Exports<'next/dist/next-server/lib/amp-context'>;
+}
+declare module 'next/dist/next-server/lib/amp.js' {
+  declare module.exports: $Exports<'next/dist/next-server/lib/amp'>;
+}
+declare module 'next/dist/next-server/lib/constants.js' {
+  declare module.exports: $Exports<'next/dist/next-server/lib/constants'>;
+}
+declare module 'next/dist/next-server/lib/data-manager-context.js' {
+  declare module.exports: $Exports<
+    'next/dist/next-server/lib/data-manager-context'
+  >;
+}
+declare module 'next/dist/next-server/lib/data-manager.js' {
+  declare module.exports: $Exports<'next/dist/next-server/lib/data-manager'>;
+}
+declare module 'next/dist/next-server/lib/document-context.js' {
+  declare module.exports: $Exports<
+    'next/dist/next-server/lib/document-context'
+  >;
+}
+declare module 'next/dist/next-server/lib/dynamic.js' {
+  declare module.exports: $Exports<'next/dist/next-server/lib/dynamic'>;
+}
+declare module 'next/dist/next-server/lib/head-manager-context.js' {
+  declare module.exports: $Exports<
+    'next/dist/next-server/lib/head-manager-context'
+  >;
+}
+declare module 'next/dist/next-server/lib/head.js' {
+  declare module.exports: $Exports<'next/dist/next-server/lib/head'>;
+}
+declare module 'next/dist/next-server/lib/loadable-context.js' {
+  declare module.exports: $Exports<
+    'next/dist/next-server/lib/loadable-context'
+  >;
+}
+declare module 'next/dist/next-server/lib/loadable.js' {
+  declare module.exports: $Exports<'next/dist/next-server/lib/loadable'>;
+}
+declare module 'next/dist/next-server/lib/mitt.js' {
+  declare module.exports: $Exports<'next/dist/next-server/lib/mitt'>;
+}
+declare module 'next/dist/next-server/lib/request-context.js' {
+  declare module.exports: $Exports<'next/dist/next-server/lib/request-context'>;
+}
+declare module 'next/dist/next-server/lib/router-context.js' {
+  declare module.exports: $Exports<'next/dist/next-server/lib/router-context'>;
+}
+declare module 'next/dist/next-server/lib/router/rewrite-url-for-export.js' {
+  declare module.exports: $Exports<
+    'next/dist/next-server/lib/router/rewrite-url-for-export'
+  >;
+}
+declare module 'next/dist/next-server/lib/router/router.js' {
+  declare module.exports: $Exports<'next/dist/next-server/lib/router/router'>;
+}
+declare module 'next/dist/next-server/lib/router/utils/index' {
+  declare module.exports: $Exports<'next/dist/next-server/lib/router/utils'>;
+}
+declare module 'next/dist/next-server/lib/router/utils/index.js' {
+  declare module.exports: $Exports<'next/dist/next-server/lib/router/utils'>;
+}
+declare module 'next/dist/next-server/lib/router/utils/is-dynamic.js' {
+  declare module.exports: $Exports<
+    'next/dist/next-server/lib/router/utils/is-dynamic'
+  >;
+}
+declare module 'next/dist/next-server/lib/router/utils/route-matcher.js' {
+  declare module.exports: $Exports<
+    'next/dist/next-server/lib/router/utils/route-matcher'
+  >;
+}
+declare module 'next/dist/next-server/lib/router/utils/route-regex.js' {
+  declare module.exports: $Exports<
+    'next/dist/next-server/lib/router/utils/route-regex'
+  >;
+}
+declare module 'next/dist/next-server/lib/router/utils/sorted-routes.js' {
+  declare module.exports: $Exports<
+    'next/dist/next-server/lib/router/utils/sorted-routes'
+  >;
+}
+declare module 'next/dist/next-server/lib/runtime-config.js' {
+  declare module.exports: $Exports<'next/dist/next-server/lib/runtime-config'>;
+}
+declare module 'next/dist/next-server/lib/side-effect.js' {
+  declare module.exports: $Exports<'next/dist/next-server/lib/side-effect'>;
+}
+declare module 'next/dist/next-server/lib/utils.js' {
+  declare module.exports: $Exports<'next/dist/next-server/lib/utils'>;
+}
+declare module 'next/dist/next-server/server/api-utils.js' {
+  declare module.exports: $Exports<'next/dist/next-server/server/api-utils'>;
+}
+declare module 'next/dist/next-server/server/config.js' {
+  declare module.exports: $Exports<'next/dist/next-server/server/config'>;
+}
+declare module 'next/dist/next-server/server/get-page-files.js' {
+  declare module.exports: $Exports<
+    'next/dist/next-server/server/get-page-files'
+  >;
+}
+declare module 'next/dist/next-server/server/lib/path-match.js' {
+  declare module.exports: $Exports<
+    'next/dist/next-server/server/lib/path-match'
+  >;
+}
+declare module 'next/dist/next-server/server/lib/recursive-readdir-sync.js' {
+  declare module.exports: $Exports<
+    'next/dist/next-server/server/lib/recursive-readdir-sync'
+  >;
+}
+declare module 'next/dist/next-server/server/load-components.js' {
+  declare module.exports: $Exports<
+    'next/dist/next-server/server/load-components'
+  >;
+}
+declare module 'next/dist/next-server/server/next-server.js' {
+  declare module.exports: $Exports<'next/dist/next-server/server/next-server'>;
+}
+declare module 'next/dist/next-server/server/normalize-page-path.js' {
+  declare module.exports: $Exports<
+    'next/dist/next-server/server/normalize-page-path'
+  >;
+}
+declare module 'next/dist/next-server/server/optimize-amp.js' {
+  declare module.exports: $Exports<'next/dist/next-server/server/optimize-amp'>;
+}
+declare module 'next/dist/next-server/server/render.js' {
+  declare module.exports: $Exports<'next/dist/next-server/server/render'>;
+}
+declare module 'next/dist/next-server/server/require.js' {
+  declare module.exports: $Exports<'next/dist/next-server/server/require'>;
+}
+declare module 'next/dist/next-server/server/router.js' {
+  declare module.exports: $Exports<'next/dist/next-server/server/router'>;
+}
+declare module 'next/dist/next-server/server/send-html.js' {
+  declare module.exports: $Exports<'next/dist/next-server/server/send-html'>;
+}
+declare module 'next/dist/next-server/server/serve-static.js' {
+  declare module.exports: $Exports<'next/dist/next-server/server/serve-static'>;
+}
+declare module 'next/dist/next-server/server/spr-cache.js' {
+  declare module.exports: $Exports<'next/dist/next-server/server/spr-cache'>;
+}
+declare module 'next/dist/next-server/server/utils.js' {
+  declare module.exports: $Exports<'next/dist/next-server/server/utils'>;
 }
 declare module 'next/dist/pages/_app.js' {
   declare module.exports: $Exports<'next/dist/pages/_app'>;
@@ -692,7 +1261,9 @@ declare module 'next/dist/server/htmlescape.js' {
   declare module.exports: $Exports<'next/dist/server/htmlescape'>;
 }
 declare module 'next/dist/server/lib/error-overlay-middleware.js' {
-  declare module.exports: $Exports<'next/dist/server/lib/error-overlay-middleware'>;
+  declare module.exports: $Exports<
+    'next/dist/server/lib/error-overlay-middleware'
+  >;
 }
 declare module 'next/dist/server/lib/find-page-file.js' {
   declare module.exports: $Exports<'next/dist/server/lib/find-page-file'>;
@@ -711,6 +1282,36 @@ declare module 'next/dist/server/next.js' {
 }
 declare module 'next/dist/server/on-demand-entry-handler.js' {
   declare module.exports: $Exports<'next/dist/server/on-demand-entry-handler'>;
+}
+declare module 'next/dist/telemetry/anonymous-meta.js' {
+  declare module.exports: $Exports<'next/dist/telemetry/anonymous-meta'>;
+}
+declare module 'next/dist/telemetry/ci-info.js' {
+  declare module.exports: $Exports<'next/dist/telemetry/ci-info'>;
+}
+declare module 'next/dist/telemetry/events/build.js' {
+  declare module.exports: $Exports<'next/dist/telemetry/events/build'>;
+}
+declare module 'next/dist/telemetry/events/index' {
+  declare module.exports: $Exports<'next/dist/telemetry/events'>;
+}
+declare module 'next/dist/telemetry/events/index.js' {
+  declare module.exports: $Exports<'next/dist/telemetry/events'>;
+}
+declare module 'next/dist/telemetry/events/plugins.js' {
+  declare module.exports: $Exports<'next/dist/telemetry/events/plugins'>;
+}
+declare module 'next/dist/telemetry/events/version.js' {
+  declare module.exports: $Exports<'next/dist/telemetry/events/version'>;
+}
+declare module 'next/dist/telemetry/post-payload.js' {
+  declare module.exports: $Exports<'next/dist/telemetry/post-payload'>;
+}
+declare module 'next/dist/telemetry/project-id.js' {
+  declare module.exports: $Exports<'next/dist/telemetry/project-id'>;
+}
+declare module 'next/dist/telemetry/storage.js' {
+  declare module.exports: $Exports<'next/dist/telemetry/storage'>;
 }
 declare module 'next/document.js' {
   declare module.exports: $Exports<'next/document'>;
