@@ -1,5 +1,5 @@
-// flow-typed signature: c6362f869480fb03bfd4618fff2f16f1
-// flow-typed version: <<STUB>>/prettier-eslint_v8.8.2/flow_v0.99.1
+// flow-typed signature: fc4ca2f03215700d2d579eb4725b9754
+// flow-typed version: <<STUB>>/prettier-eslint_v9.0.0/flow_v0.112.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -22,7 +22,7 @@ declare module 'prettier-eslint' {
  * require those files directly. Feel free to delete any files that aren't
  * needed.
  */
-declare module 'prettier-eslint/dist/index' {
+declare module 'prettier-eslint/dist' {
   declare module.exports: any;
 }
 
@@ -31,8 +31,11 @@ declare module 'prettier-eslint/dist/utils' {
 }
 
 // Filename aliases
+declare module 'prettier-eslint/dist/index' {
+  declare module.exports: $Exports<'prettier-eslint/dist'>;
+}
 declare module 'prettier-eslint/dist/index.js' {
-  declare module.exports: $Exports<'prettier-eslint/dist/index'>;
+  declare module.exports: $Exports<'prettier-eslint/dist'>;
 }
 declare module 'prettier-eslint/dist/utils.js' {
   declare module.exports: $Exports<'prettier-eslint/dist/utils'>;
