@@ -1,5 +1,5 @@
-// flow-typed signature: aed9b3bbd2f1c41ae273894dc2e5dbee
-// flow-typed version: <<STUB>>/sw-precache-webpack-plugin_v0.11.5/flow_v0.99.1
+// flow-typed signature: 435c7e4b99d137ec2e364cb6a356f270
+// flow-typed version: <<STUB>>/sw-precache-webpack-plugin_v0.11.5/flow_v0.112.0
 
 /**
  * This is an autogenerated libdef stub for:
@@ -22,18 +22,24 @@ declare module 'sw-precache-webpack-plugin' {
  * require those files directly. Feel free to delete any files that aren't
  * needed.
  */
-declare module 'sw-precache-webpack-plugin/lib/index' {
+declare module 'sw-precache-webpack-plugin/lib' {
   declare module.exports: any;
 }
 
-declare module 'sw-precache-webpack-plugin/src/index' {
+declare module 'sw-precache-webpack-plugin/src' {
   declare module.exports: any;
 }
 
 // Filename aliases
+declare module 'sw-precache-webpack-plugin/lib/index' {
+  declare module.exports: $Exports<'sw-precache-webpack-plugin/lib'>;
+}
 declare module 'sw-precache-webpack-plugin/lib/index.js' {
-  declare module.exports: $Exports<'sw-precache-webpack-plugin/lib/index'>;
+  declare module.exports: $Exports<'sw-precache-webpack-plugin/lib'>;
+}
+declare module 'sw-precache-webpack-plugin/src/index' {
+  declare module.exports: $Exports<'sw-precache-webpack-plugin/src'>;
 }
 declare module 'sw-precache-webpack-plugin/src/index.js' {
-  declare module.exports: $Exports<'sw-precache-webpack-plugin/src/index'>;
+  declare module.exports: $Exports<'sw-precache-webpack-plugin/src'>;
 }
